refactor(http): split into submodules (re-#625)
Split the 2,497-line crates/harness-server/src/http.rs into focused
submodules under src/http/, each under 800 lines. Zero behavior change;
public API preserved via pub use re-exports.

Modules extracted:
- types.rs       — AppState and service structs
- handlers.rs    — HTTP handler functions
- intake_auth.rs — signal ingestion and auth handlers
- init.rs        — build_app_state and completion callback
- startup_recovery.rs — redispatch helpers on startup
- server.rs      — serve() and build_router()
- helpers.rs     — parse_pr_num_from_url, resolve_reviewer
- builders.rs    — task builder helpers
- rate_limit.rs  — SignalRateLimiter
- auth.rs        — API auth middleware
- task_routes.rs — task CRUD routes

Refs #625
