+++ crates/harness-server/src/http.rs
@@ -1,2497 +1,26 @@
-use crate::{router, server::HarnessServer, task_runner};
-use axum::{
-    body::Bytes,
-    extract::DefaultBodyLimit,
-    extract::{Path, State},
-    http::{HeaderMap, StatusCode},
-    middleware::{self},
-    response::{
-        sse::{Event, KeepAlive, Sse},
-        IntoResponse, Response,
-    },
-    routing::{get, post},
-    Json, Router,
-};
-use dashmap::DashMap;
-use harness_protocol::{methods::RpcRequest, notifications::RpcNotification};
-use serde_json::json;
-use std::net::SocketAddr;
-use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-use std::sync::Arc;
-use tokio::sync::{broadcast, Mutex, RwLock};
-
 pub(crate) mod auth;
 pub(crate) mod builders;
+pub(crate) mod handlers;
+pub(crate) mod helpers;
+pub(crate) mod init;
+pub(crate) mod intake_auth;
 pub(crate) mod rate_limit;
+pub(crate) mod server;
+pub(crate) mod startup_recovery;
 pub(crate) mod task_routes;
-
-/// Core services: thread/task management and persistence.
-pub struct CoreServices {
-    pub server: Arc<HarnessServer>,
-    pub project_root: std::path::PathBuf,
-    /// Home directory captured at startup to avoid TOCTOU when validating
-    /// project roots against `$HOME` in concurrent requests.
-    pub home_dir: std::path::PathBuf,
-    pub tasks: Arc<task_runner::TaskStore>,
-    pub thread_db: Option<crate::thread_db::ThreadDb>,
-    pub plan_db: Option<crate::plan_db::PlanDb>,
-    /// In-memory plan cache hydrated from `plan_db` on startup.
-    /// Write-through: every mutation must also persist via `plan_db`.
-    pub plan_cache: Arc<DashMap<String, harness_exec::plan::ExecPlan>>,
-    pub project_registry: Option<std::sync::Arc<crate::project_registry::ProjectRegistry>>,
-    pub runtime_state_store: Option<Arc<crate::runtime_state_store::RuntimeStateStore>>,
-    /// Q-value store for MemRL rule utility tracking. None when unavailable.
-    pub q_values: Option<Arc<crate::q_value_store::QValueStore>>,
-}
-
-/// Engine services: skills, rules, and garbage collection.
-pub struct EngineServices {
-    pub skills: Arc<RwLock<harness_skills::store::SkillStore>>,
-    pub rules: Arc<RwLock<harness_rules::engine::RuleEngine>>,
-    pub gc_agent: Arc<harness_gc::gc_agent::GcAgent>,
-}
-
-/// Observability services: event store and telemetry.
-pub struct ObservabilityServices {
-    pub events: Arc<harness_observe::event_store::EventStore>,
-    pub signal_rate_limiter: Arc<rate_limit::SignalRateLimiter>,
-    pub password_reset_rate_limiter: Arc<rate_limit::PasswordResetRateLimiter>,
-    pub review_store: Option<Arc<crate::review_store::ReviewStore>>,
-}
-
-/// Concurrency services: task queue and workspace isolation.
-pub struct ConcurrencyServices {
-    pub task_queue: Arc<crate::task_queue::TaskQueue>,
-    pub workspace_mgr: Option<Arc<crate::workspace::WorkspaceManager>>,
-}
-
-/// Notification services: broadcast channels and lag tracking.
-pub struct NotificationServices {
-    /// Broadcast channel for server-push notifications (WebSocket and stdio transports).
-    pub notification_tx: broadcast::Sender<RpcNotification>,
-    /// Total number of dropped broadcast notifications due to lagged receivers.
-    pub notification_lagged_total: Arc<AtomicU64>,
-    /// Log lagged drops when the total crosses a multiple of this value.
-    /// Set to 0 to disable lag logs while still counting drops.
-    pub notification_lag_log_every: u64,
-    /// Channel for server-push JSON-RPC notifications (stdio transport only).
-    pub notify_tx: Option<crate::notify::NotifySender>,
-    /// Whether `initialize` has been received (but `initialized` may not yet be set).
-    /// Used to enforce the `initialize` → `initialized` ordering.
-    pub initializing: Arc<AtomicBool>,
-    /// Whether the client has completed the initialize/initialized handshake.
-    pub initialized: Arc<AtomicBool>,
-    /// Broadcast channel used to signal all active WebSocket connections to close gracefully.
-    pub ws_shutdown_tx: broadcast::Sender<()>,
-}
-
-/// Intake services: external event sources and task completion handling.
-pub struct IntakeServices {
-    /// Feishu Bot intake handler. None when feishu intake is disabled or not configured.
-    pub feishu_intake: Option<Arc<crate::intake::feishu::FeishuIntake>>,
-    /// All GitHub issue pollers, one per configured repo. The same Arc instances
-    /// are shared between the completion callback and the orchestrator so that
-    /// `on_task_complete` (e.g. evicting a failed issue from `dispatched`)
-    /// operates on the live poller rather than a detached clone.
-    pub github_pollers: Vec<Arc<dyn crate::intake::IntakeSource>>,
-    /// Completion callback invoked when a task reaches a terminal state.
-    pub completion_callback: Option<task_runner::CompletionCallback>,
-}
-
-pub struct AppState {
-    pub core: CoreServices,
-    pub engines: EngineServices,
-    pub observability: ObservabilityServices,
-    pub concurrency: ConcurrencyServices,
-    pub runtime_hosts: Arc<crate::runtime_hosts::RuntimeHostManager>,
-    pub runtime_project_cache: Arc<crate::runtime_project_cache::RuntimeProjectCacheManager>,
-    /// Serializes runtime snapshot writes to avoid out-of-order persistence.
-    pub runtime_state_persist_lock: Mutex<()>,
-    /// Set when a runtime-state persist fails; the next successful
-    /// `persist_runtime_state` call clears it.  Handlers that find no
-    /// in-memory mutation (e.g. idempotent deregister retry) still trigger a
-    /// persist when this flag is set, converging durable state.
-    pub runtime_state_dirty: AtomicBool,
-    pub notifications: NotificationServices,
-    pub intake: IntakeServices,
-    pub interceptors: Vec<Arc<dyn harness_core::interceptor::TurnInterceptor>>,
-
-    // ── Service layer ────────────────────────────────────────────────────────
-    // Trait-based abstractions for independent testability. Each service owns
-    // its dependencies; the fields above are preserved for handlers that have
-    // not yet been migrated to the service interfaces.
-    /// Project registry operations and default-root lookup.
-    pub project_svc: Arc<dyn crate::services::project::ProjectService>,
-    /// Task lifecycle queries and stream subscriptions.
-    pub task_svc: Arc<dyn crate::services::task::TaskService>,
-    /// Task enqueue: project resolution, agent dispatch, concurrency, workspace.
-    pub execution_svc: Arc<dyn crate::services::execution::ExecutionService>,
-}
-
-impl AppState {
-    pub fn observe_notification_lag(&self, dropped: u64) -> u64 {
-        let previous_total = self
-            .notifications
-            .notification_lagged_total
-            .fetch_add(dropped, Ordering::Relaxed);
-        let dropped_total = previous_total.saturating_add(dropped);
-        let log_every = self.notifications.notification_lag_log_every;
-        if log_every > 0 && previous_total / log_every < dropped_total / log_every {
-            tracing::warn!(
-                dropped_since_last_recv = dropped,
-                dropped_total,
-                log_every,
-                "notification receiver lagged; dropped broadcast notifications"
-            );
-        }
-        dropped_total
-    }
-
-    pub async fn persist_runtime_state(&self) -> anyhow::Result<()> {
-        let _guard = self.runtime_state_persist_lock.lock().await;
-        let Some(store) = self.core.runtime_state_store.as_ref() else {
-            return Ok(());
-        };
-        let (hosts, leases) = self.runtime_hosts.snapshot_state();
-        let project_caches = self.runtime_project_cache.snapshot_state();
-        match store.persist_snapshot(hosts, leases, project_caches).await {
-            Ok(()) => {
-                self.runtime_state_dirty.store(false, Ordering::Release);
-                Ok(())
-            }
-            Err(e) => {
-                self.runtime_state_dirty.store(true, Ordering::Release);
-                Err(e)
-            }
-        }
-    }
-
-    /// Returns `true` when a previous persist failed and durable state may be
-    /// stale.  Handlers that skip their own mutation (e.g. idempotent
-    /// deregister retry returning NOT_FOUND) should check this and re-persist
-    /// to converge.
-    pub fn is_runtime_state_dirty(&self) -> bool {
-        self.runtime_state_dirty.load(Ordering::Acquire)
-    }
-}
-fn resolve_project_root(configured_root: &std::path::Path) -> anyhow::Result<std::path::PathBuf> {
-    let project_root = configured_root.canonicalize().map_err(|e| {
-        anyhow::anyhow!(
-            "invalid server.project_root '{}': {e}",
-            configured_root.display()
-        )
-    })?;
-    if !project_root.is_dir() {
-        anyhow::bail!(
-            "server.project_root is not a directory: {}",
-            project_root.display()
-        );
-    }
-    Ok(project_root)
-}
-
-/// Expand a leading `~/` or standalone `~` to the value of `$HOME`.
-/// Returns the path unchanged when `~` is not present or `HOME` is unset.
-fn expand_tilde(path: &std::path::Path) -> std::path::PathBuf {
-    if let Some(s) = path.to_str() {
-        if let Some(rest) = s.strip_prefix("~/") {
-            if let Ok(home) = std::env::var("HOME") {
-                return std::path::PathBuf::from(home).join(rest);
-            }
-        } else if s == "~" {
-            if let Ok(home) = std::env::var("HOME") {
-                return std::path::PathBuf::from(home);
-            }
-        }
-    }
-    path.to_path_buf()
-}
-
-/// Build an AppState with all stores. Used by both HTTP and stdio transports.
-///
-/// Initialization is split into five ordered phases — each phase's outputs
-/// become inputs to the next.  Dependency edges:
-///   storage → engines → registry (needs storage.tasks for orphan cleanup)
-///   registry → intake  (needs engines.gc_agent + events, registry.project_registry)
-///   intake   → services (needs interceptors wired to engines.rules + events)
-pub async fn build_app_state(server: Arc<HarnessServer>) -> anyhow::Result<AppState> {
-    let dir = expand_tilde(&server.config.server.data_dir);
-    let project_root = resolve_project_root(&server.config.server.project_root)?;
-
-    tracing::debug!(
-        data_dir = %dir.display(),
-        project_root = %project_root.display(),
-        discovery_paths = ?server.config.rules.discovery_paths,
-        builtin_path = ?server.config.rules.builtin_path,
-        exec_policy_paths = ?server.config.rules.exec_policy_paths,
-        requirements_path = ?server.config.rules.requirements_path,
-        session_renewal_secs = server.config.observe.session_renewal_secs,
-        log_retention_days = server.config.observe.log_retention_days,
-        "config details (use RUST_LOG=debug to see)"
-    );
-    match server.config.server.github_webhook_secret.as_deref() {
-        Some("") => {
-            tracing::warn!(
-                "server.github_webhook_secret is configured as empty string; refusing webhook requests until this is set to a non-empty value"
-            );
-        }
-        None => {
-            tracing::warn!(
-                "server.github_webhook_secret is not configured; refusing webhook requests until this is set to a non-empty value"
-            );
-        }
-        Some(_) => {}
-    }
-
-    // Phase 1: storage — dir validation (symlink check, chmod), task DB, q_value DB.
-    let storage = builders::storage::build_storage(&dir).await?;
-
-    // Phase 2: engines — rule engine, event store (+purge task), GC agent, skill store.
-    // Depends on: storage (none directly, but must precede registry which uses storage.tasks).
-    let engines = builders::engines::build_engines(&server, &dir, &project_root).await?;
-
-    // Phase 3: registry — thread DB, plan DB + cache, project registry, workspace manager,
-    // runtime state store.
-    // Depends on: storage.tasks (orphan-worktree cleanup reads terminal task IDs).
-    let registry =
-        builders::registry::build_registry(&server, &dir, &project_root, &storage.tasks).await?;
-
-    // Phase 4: intake — task queue, Feishu/GitHub pollers, quality trigger, completion callback
-    // (including Q-value wrapper).
-    // Depends on: engines.gc_agent + engines.events (quality trigger),
-    //             registry.project_registry (unused directly but ordering is stable).
-    let intake =
-        builders::intake::build_intake(&server, &storage, &engines, &registry, &project_root, &dir)
-            .await?;
-
-    // Phase 5: services — interceptor stack, service layer, runtime host/project-cache
-    // managers, snapshot restore, recovery validator spawn.
-    // Depends on: engines.rules + engines.events (interceptors),
-    //             registry.workspace_mgr + registry.project_registry (execution service),
-    //             intake.task_queue + intake.completion_callback (execution service).
-    let services = builders::services::build_services(
-        &server,
-        &storage,
-        &engines,
-        &registry,
-        &intake,
-        &project_root,
-    )
-    .await?;
-
-    let configured_capacity = server.config.server.notification_broadcast_capacity;
-    let notification_broadcast_capacity = configured_capacity.max(1);
-    let notification_lag_log_every = server.config.server.notification_lag_log_every;
-    if configured_capacity == 0 {
-        tracing::warn!(
-            "server.notification_broadcast_capacity=0 is invalid; falling back to capacity=1"
-        );
-    }
-
-    let signal_rate_limit = server.config.server.signal_rate_limit_per_minute;
-    let password_reset_rate_limit = server.config.server.password_reset_rate_limit_per_hour;
-    let home_dir = std::env::var("HOME")
-        .map(std::path::PathBuf::from)
-        .unwrap_or_else(|_| project_root.clone());
-
-    Ok(AppState {
-        core: CoreServices {
-            server,
-            project_root,
-            home_dir,
-            tasks: storage.tasks,
-            thread_db: Some(registry.thread_db),
-            plan_db: Some(registry.plan_db),
-            plan_cache: registry.plan_cache,
-            project_registry: Some(registry.project_registry),
-            runtime_state_store: registry.runtime_state_store,
-            q_values: storage.q_values,
-        },
-        engines: EngineServices {
-            skills: engines.skills,
-            rules: engines.rules,
-            gc_agent: engines.gc_agent,
-        },
-        observability: ObservabilityServices {
-            events: engines.events,
-            signal_rate_limiter: Arc::new(rate_limit::SignalRateLimiter::new(signal_rate_limit)),
-            password_reset_rate_limiter: Arc::new(rate_limit::PasswordResetRateLimiter::new(
-                password_reset_rate_limit,
-            )),
-            review_store: {
-                let review_db_path = harness_core::config::dirs::default_db_path(&dir, "reviews");
-                match crate::review_store::ReviewStore::open(&review_db_path).await {
-                    Ok(store) => Some(Arc::new(store)),
-                    Err(e) => {
-                        tracing::warn!(
-                            "review store init failed, reviews will not be persisted: {e}"
-                        );
-                        None
-                    }
-                }
-            },
-        },
-        concurrency: ConcurrencyServices {
-            task_queue: intake.task_queue,
-            workspace_mgr: registry.workspace_mgr,
-        },
-        runtime_hosts: services.runtime_hosts,
-        runtime_project_cache: services.runtime_project_cache,
-        runtime_state_persist_lock: Mutex::new(()),
-        runtime_state_dirty: AtomicBool::new(false),
-        notifications: NotificationServices {
-            notification_tx: broadcast::channel(notification_broadcast_capacity).0,
-            notification_lagged_total: Arc::new(AtomicU64::new(0)),
-            notification_lag_log_every,
-            notify_tx: None,
-            initializing: Arc::new(AtomicBool::new(false)),
-            initialized: Arc::new(AtomicBool::new(false)),
-            ws_shutdown_tx: broadcast::channel(1).0,
-        },
-        interceptors: services.interceptors,
-        intake: IntakeServices {
-            feishu_intake: intake.feishu_intake,
-            github_pollers: intake.github_pollers.into_iter().map(|(_, p)| p).collect(),
-            completion_callback: intake.completion_callback,
-        },
-        project_svc: services.project_svc,
-        task_svc: services.task_svc,
-        execution_svc: services.execution_svc,
-    })
-}
-
-pub(crate) fn build_completion_callback(
-    feishu_intake: &Option<Arc<crate::intake::feishu::FeishuIntake>>,
-    github_pollers: &[(String, Arc<dyn crate::intake::IntakeSource>)],
-    review_config: harness_core::config::agents::AgentReviewConfig,
-    quality_trigger: Option<Arc<crate::quality_trigger::QualityTrigger>>,
-    config_github_token: Option<String>,
-) -> Option<task_runner::CompletionCallback> {
-    let mut sources: std::collections::HashMap<String, Arc<dyn crate::intake::IntakeSource>> =
-        std::collections::HashMap::new();
-    // Insert each GitHub poller keyed by "github:{owner/repo}" for precise
-    // per-repo routing. Also insert the first poller under the bare "github"
-    // key as a backward-compat fallback for tasks that pre-date multi-repo
-    // support and have task.repo == None.
-    for (i, (key, poller)) in github_pollers.iter().enumerate() {
-        sources.insert(key.clone(), poller.clone());
-        if i == 0 {
-            sources.insert("github".to_string(), poller.clone());
-        }
-    }
-    if let Some(fi) = feishu_intake {
-        let fi_source: Arc<dyn crate::intake::IntakeSource> = fi.clone();
-        sources.insert(fi_source.name().to_string(), fi_source);
-    }
-    if sources.is_empty() && !review_config.review_bot_auto_trigger && quality_trigger.is_none() {
-        return None;
-    }
-    let sources = Arc::new(sources);
-    Some(Arc::new(move |task: task_runner::TaskState| {
-        let sources = sources.clone();
-        let review_config = review_config.clone();
-        let quality_trigger = quality_trigger.clone();
-        let github_token = config_github_token.clone();
-        Box::pin(async move {
-            // Grade recent events and auto-trigger GC if quality is poor.
-            if let Some(qt) = quality_trigger {
-                let task_ctx = task.pr_url.as_ref().and_then(|pr| {
-                    // Find the most recent implementation-affecting round.
-                    // agent_review_fix rounds always store detail: None, so if
-                    // the most recent such round is a fix round we have no
-                    // usable diff for the final PR state — skip cross-review
-                    // entirely to avoid judging stale initial-implement content
-                    // and spuriously downgrading an already-fixed PR.
-                    let last_impl_round = task
-                        .rounds
-                        .iter()
-                        .rev()
-                        .find(|r| r.action == "implement" || r.action == "agent_review_fix");
-                    let diff = match last_impl_round {
-                        Some(r) if r.action == "implement" => {
-                            // Use the full implement-agent output as review context.
-                            // The implementation prompt contract requires a PR_URL line
-                            // but does not mandate unified-diff format, so accepting
-                            // any non-empty detail avoids silently dropping valid rounds.
-                            r.detail.clone().unwrap_or_default()
-                        }
-                        // agent_review_fix (detail always None) or no round at all
-                        _ => String::new(),
-                    };
-                    if diff.is_empty() {
-                        None
-                    } else {
-                        Some(crate::quality_trigger::TaskReviewContext {
-                            diff,
-                            pr_description: pr.clone(),
-                        })
-                    }
-                });
-                qt.check_and_maybe_trigger(task_ctx.as_ref()).await;
-            }
-
-            // Auto-trigger review bot comment when task completes with a PR URL.
-            if review_config.review_bot_auto_trigger {
-                if let task_runner::TaskStatus::Done = &task.status {
-                    if let Some(pr_url) = task.pr_url.as_deref() {
-                        if let Some((owner, repo, pr_num)) =
-                            harness_core::prompts::parse_github_pr_url(pr_url)
-                        {
-                            let resolved_token = github_token
-                                .or_else(|| std::env::var("GITHUB_TOKEN").ok())
-                                .filter(|t| !t.is_empty());
-                            match resolved_token {
-                                Some(token) => {
-                                    if let Err(e) = post_review_bot_comment(
-                                        &owner,
-                                        &repo,
-                                        pr_num,
-                                        &review_config.review_bot_command,
-                                        &token,
-                                    )
-                                    .await
-                                    {
-                                        tracing::warn!(
-                                            pr_url,
-                                            "review_bot_auto_trigger: failed to post comment: {e}"
-                                        );
-                                    } else {
-                                        tracing::info!(
-                                            pr_url,
-                                            comment = review_config.review_bot_command,
-                                            "review bot comment posted"
-                                        );
-                                    }
-                                }
-                                _ => {
-                                    tracing::warn!(
-                                        pr_url,
-                                        "review_bot_auto_trigger: GITHUB_TOKEN not set or empty; skipping"
-                                    );
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            // Intake source notification.
-            let Some(source_name) = task.source.as_deref() else {
-                return;
-            };
-            let Some(external_id) = task.external_id.as_deref() else {
-                tracing::warn!(
-                    task_id = ?task.id,
-                    source = source_name,
-                    "completion_callback: task missing external_id, skipping"
-                );
-                return;
-            };
-            // For GitHub tasks, route to the specific repo's poller using
-            // "github:{owner/repo}". Fall back to the bare "github" key for
-            // tasks persisted before multi-repo support (task.repo == None).
-            let lookup_key = if source_name == "github" {
-                task.repo
-                    .as_ref()
-                    .map(|r| format!("github:{r}"))
-                    .unwrap_or_else(|| "github".to_string())
-            } else {
-                source_name.to_string()
-            };
-            let Some(source) = sources.get(&lookup_key) else {
-                tracing::warn!(
-                    task_id = ?task.id,
-                    source = source_name,
-                    lookup_key,
-                    "completion_callback: intake source not found, skipping"
-                );
-                return;
-            };
-            let summary = if task.status.is_success() {
-                task.pr_url
-                    .as_deref()
-                    .map(|url| format!("PR: {url}"))
-                    .unwrap_or_else(|| "Task completed.".to_string())
-            } else if task.status.is_cancelled() {
-                "Task cancelled.".to_string()
-            } else if task.status.is_failure() {
-                task.error.as_deref().unwrap_or("unknown error").to_string()
-            } else {
-                tracing::warn!(
-                    task_id = ?task.id,
-                    status = ?task.status,
-                    "completion_callback: called with non-terminal status, skipping"
-                );
-                return;
-            };
-            let result = crate::intake::TaskCompletionResult {
-                status: task.status.clone(),
-                pr_url: task.pr_url.clone(),
-                error: task.error.clone(),
-                summary,
-            };
-            if let Err(e) = source.on_task_complete(external_id, &result).await {
-                tracing::warn!(
-                    task_id = ?task.id,
-                    source = source_name,
-                    "on_task_complete failed: {e}"
-                );
-            }
-        })
-    }))
-}
-
-/// Post a comment to a GitHub PR via the Issues API.
-async fn post_review_bot_comment(
-    owner: &str,
-    repo: &str,
-    pr_number: u64,
-    body: &str,
-    github_token: &str,
-) -> anyhow::Result<()> {
-    let url = format!("https://api.github.com/repos/{owner}/{repo}/issues/{pr_number}/comments");
-    let client = reqwest::Client::new();
-    let resp = client
-        .post(&url)
-        .header("Authorization", format!("Bearer {github_token}"))
-        .header("Accept", "application/vnd.github+json")
-        .header("X-GitHub-Api-Version", "2022-11-28")
-        .header("User-Agent", "harness-bot")
-        .json(&serde_json::json!({ "body": body }))
-        .send()
-        .await?;
-    if !resp.status().is_success() {
-        let status = resp.status();
-        let text = resp.text().await.unwrap_or_default();
-        anyhow::bail!("GitHub API returned {status}: {text}");
-    }
-    Ok(())
-}
-
-/// Resolve the reviewer agent for independent agent review.
-///
-/// 1. If `config.reviewer_agent` is set and differs from implementor, use it.
-/// 2. Otherwise, auto-select the first registered agent that isn't the implementor.
-/// 3. If none found, return None (agent review will be skipped).
-pub(crate) fn resolve_reviewer(
-    registry: &harness_agents::registry::AgentRegistry,
-    config: &harness_core::config::agents::AgentReviewConfig,
-    implementor_name: &str,
-) -> (
-    Option<Arc<dyn harness_core::agent::CodeAgent>>,
-    harness_core::config::agents::AgentReviewConfig,
-) {
-    if !config.enabled {
-        return (None, config.clone());
-    }
-
-    // Explicit reviewer
-    if !config.reviewer_agent.is_empty() {
-        if config.reviewer_agent == implementor_name {
-            tracing::warn!(
-                "agents.review.reviewer_agent == implementor '{}', skipping agent review",
-                implementor_name
-            );
-            return (None, config.clone());
-        }
-        if let Some(agent) = registry.get(&config.reviewer_agent) {
-            return (Some(agent), config.clone());
-        }
-        tracing::warn!(
-            "agents.review.reviewer_agent '{}' not registered, skipping agent review",
-            config.reviewer_agent
-        );
-        return (None, config.clone());
-    }
-
-    // Auto-select: first agent != implementor
-    for name in registry.list() {
-        if name != implementor_name {
-            if let Some(agent) = registry.get(name) {
-                return (Some(agent), config.clone());
-            }
-        }
-    }
-
-    (None, config.clone())
-}
-
-/// Extract the PR number from a GitHub PR URL.
-///
-/// Handles:
-/// - `.../pull/42`
-/// - `.../pull/42/files`
-/// - `.../pull/42#discussion_r...`
-pub(crate) fn parse_pr_num_from_url(url: &str) -> Option<u64> {
-    // Strip fragment first, then query string
-    let url = url.split('#').next().unwrap_or(url);
-    let url = url.split('?').next().unwrap_or(url);
-    // Walk path segments looking for "pull", then parse the segment that follows
-    let mut parts = url.split('/');
-    while let Some(seg) = parts.next() {
-        if seg == "pull" {
-            return parts.next()?.parse::<u64>().ok();
-        }
-    }
-    None
-}
-
-pub async fn serve(server: Arc<HarnessServer>, addr: SocketAddr) -> anyhow::Result<()> {
-    tracing::info!("harness: HTTP server listening on {addr}");
-    // Record true server start time before accepting any connections.
-    crate::handlers::dashboard::SERVER_START.get_or_init(std::time::Instant::now);
-
-    let state = Arc::new(build_app_state(server.clone()).await?);
-
-    // Startup summary — one clean line instead of scattered logs.
-    {
-        let guard_count = state.engines.rules.read().await.guards().len();
-        let skill_count = state.engines.skills.read().await.list().len();
-        let task_count = state.core.tasks.list_all().len();
-        tracing::info!(
-            project = %state.core.project_root.display(),
-            guards = guard_count,
-            skills = skill_count,
-            pending_tasks = task_count,
-            "harness: ready"
-        );
-    }
-
-    // Spawn background watcher for AwaitingDeps tasks.
-    // Uses Weak<AppState> to avoid a reference cycle; the loop exits when AppState is dropped.
-    {
-        let weak_state = Arc::downgrade(&state);
-        tokio::spawn(async move {
-            let mut interval = tokio::time::interval(tokio::time::Duration::from_secs(10));
-            interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-            loop {
-                interval.tick().await;
-                let state = match weak_state.upgrade() {
-                    Some(s) => s,
-                    None => break,
-                };
-                let (ready_ids, failed_ids) =
-                    crate::task_runner::check_awaiting_deps(&state.core.tasks).await;
-                // Persist status changes for both ready and failed tasks.
-                for task_id in ready_ids.iter().chain(failed_ids.iter()) {
-                    if let Err(e) = state.core.tasks.persist(task_id).await {
-                        tracing::warn!(
-                            "dep-watcher: failed to persist {} after transition: {e}",
-                            task_id.0
-                        );
-                    }
-                }
-                // Invoke completion_callback for dep-failed tasks so intake
-                // sources (e.g. github_issues) can unmark them from the
-                // dispatched map and allow retries on the next poll cycle.
-                // Without this the issue stays permanently "dispatched" and
-                // the intake poller never re-queues it.
-                if let Some(ref cb) = state.intake.completion_callback {
-                    for task_id in &failed_ids {
-                        if let Some(task) = state.core.tasks.get(task_id) {
-                            let cb = cb.clone();
-                            tokio::spawn(async move {
-                                cb(task).await;
-                            });
-                        } else {
-                            tracing::warn!(
-                                "dep-watcher: task {} not found after dep-failed transition",
-                                task_id.0
-                            );
-                        }
-                    }
-                }
-                // Spawn an agent for each task whose deps are now satisfied.
-                for task_id in ready_ids {
-                    let state = state.clone();
-                    tokio::spawn(async move {
-                        let task = match state.core.tasks.get(&task_id) {
-                            Some(t) => t,
-                            None => {
-                                tracing::warn!(
-                                    "dep-watcher: task {} not found after Pending transition",
-                                    task_id.0
-                                );
-                                return;
-                            }
-                        };
-                        // Reconstruct the project path: prefer stored project_root,
-                        // fall back to repo slug so the agent runs in the right worktree.
-                        let project_path = task
-                            .project_root
-                            .clone()
-                            .or_else(|| task.repo.as_deref().map(std::path::PathBuf::from));
-                        let canonical =
-                            match task_runner::resolve_canonical_project(project_path).await {
-                                Ok(c) => c,
-                                Err(e) => {
-                                    let reason =
-                                        format!("dep-watcher: failed to resolve project path: {e}");
-                                    tracing::error!(task_id = ?task.id, "{reason}");
-                                    if let Err(pe) = task_runner::mutate_and_persist(
-                                        &state.core.tasks,
-                                        &task.id,
-                                        move |s| {
-                                            s.status = task_runner::TaskStatus::Failed;
-                                            s.error = Some(reason);
-                                        },
-                                    )
-                                    .await
-                                    {
-                                        tracing::error!(
-                                            task_id = ?task.id,
-                                            "dep-watcher: failed to persist failed status: {pe}"
-                                        );
-                                    }
-                                    return;
-                                }
-                            };
-                        let project_id = canonical.to_string_lossy().into_owned();
-
-                        // Reconstruct the CreateTaskRequest from persisted task fields.
-                        // Parse issue/pr numbers from the canonical external_id (e.g. "issue:42").
-                        let (issue, pr) = task
-                            .external_id
-                            .as_deref()
-                            .map(|eid| {
-                                if let Some(n) = eid.strip_prefix("issue:") {
-                                    (n.parse::<u64>().ok(), None)
-                                } else if let Some(n) = eid.strip_prefix("pr:") {
-                                    (None, n.parse::<u64>().ok())
-                                } else {
-                                    (None, None)
-                                }
-                            })
-                            .unwrap_or((None, None));
-                        let mut req = task_runner::CreateTaskRequest {
-                            issue,
-                            pr,
-                            project: Some(canonical),
-                            repo: task.repo.clone(),
-                            source: task.source.clone(),
-                            external_id: task.external_id.clone(),
-                            parent_task_id: task.parent_id.clone(),
-                            priority: task.priority,
-                            ..Default::default()
-                        };
-                        // Restore execution limits and prompt from persisted settings.
-                        if let Some(ref settings) = task.request_settings {
-                            settings.apply_to_req(&mut req);
-                        }
-
-                        // Guard: prompt-only tasks store their prompt in memory only
-                        // (#[serde(skip)]). After a server restart the prompt field is
-                        // absent. If no issue/pr is present either, dispatching would
-                        // call implement_from_prompt("") — a silent mis-execution.
-                        // Fail the task explicitly so the caller can re-submit.
-                        if req.prompt.is_none() && req.issue.is_none() && req.pr.is_none() {
-                            let reason = "dep-watcher: prompt-only task has no persisted \
-                                prompt after server restart; please re-submit the task"
-                                .to_string();
-                            tracing::error!(task_id = ?task.id, "{reason}");
-                            if let Err(pe) = task_runner::mutate_and_persist(
-                                &state.core.tasks,
-                                &task.id,
-                                move |s| {
-                                    s.status = task_runner::TaskStatus::Failed;
-                                    s.error = Some(reason);
-                                },
-                            )
-                            .await
-                            {
-                                tracing::error!(
-                                    task_id = ?task.id,
-                                    "dep-watcher: failed to persist failed status: {pe}"
-                                );
-                            }
-                            // Fire completion callback so intake sources (e.g. GitHub Issues
-                            // poller) remove this task from their `dispatched` map. Without
-                            // this the issue stays marked as dispatched forever and will never
-                            // be re-queued, causing a silent production deadlock.
-                            if let Some(cb) = &state.intake.completion_callback {
-                                if let Some(final_state) = state.core.tasks.get(&task.id) {
-                                    cb(final_state).await;
-                                }
-                            }
-                            return;
-                        }
-
-                        let permit = match state
-                            .concurrency
-                            .task_queue
-                            .acquire(&project_id, task.priority)
-                            .await
-                        {
-                            Ok(p) => p,
-                            Err(e) => {
-                                let reason = format!(
-                                    "dep-watcher: failed to acquire concurrency permit: {e}"
-                                );
-                                tracing::error!(task_id = ?task.id, "{reason}");
-                                if let Err(pe) = task_runner::mutate_and_persist(
-                                    &state.core.tasks,
-                                    &task.id,
-                                    move |s| {
-                                        s.status = task_runner::TaskStatus::Failed;
-                                        s.error = Some(reason);
-                                    },
-                                )
-                                .await
-                                {
-                                    tracing::error!(
-                                        task_id = ?task.id,
-                                        "dep-watcher: failed to persist failed status: {pe}"
-                                    );
-                                }
-                                return;
-                            }
-                        };
-
-                        let agent = match task_routes::select_agent(
-                            &req,
-                            &state.core.server.agent_registry,
-                            None,
-                        ) {
-                            Ok(a) => a,
-                            Err(e) => {
-                                let reason = format!("dep-watcher: failed to select agent: {e}");
-                                tracing::error!(task_id = ?task.id, "{reason}");
-                                if let Err(pe) = task_runner::mutate_and_persist(
-                                    &state.core.tasks,
-                                    &task.id,
-                                    move |s| {
-                                        s.status = task_runner::TaskStatus::Failed;
-                                        s.error = Some(reason);
-                                    },
-                                )
-                                .await
-                                {
-                                    tracing::error!(
-                                        task_id = ?task.id,
-                                        "dep-watcher: failed to persist failed status: {pe}"
-                                    );
-                                }
-                                return;
-                            }
-                        };
-                        let (reviewer, _) = resolve_reviewer(
-                            &state.core.server.agent_registry,
-                            &state.core.server.config.agents.review,
-                            agent.name(),
-                        );
-                        state.core.tasks.register_task_stream(&task.id);
-                        task_runner::spawn_preregistered_task(
-                            task.id,
-                            state.core.tasks.clone(),
-                            agent,
-                            reviewer,
-                            Arc::new(state.core.server.config.clone()),
-                            state.engines.skills.clone(),
-                            state.observability.events.clone(),
-                            state.interceptors.clone(),
-                            req,
-                            state.concurrency.workspace_mgr.clone(),
-                            permit,
-                            state.intake.completion_callback.clone(),
-                            None,
-                        )
-                        .await;
-                    });
-                }
-            }
-        });
-    }
-
-    // Re-dispatch tasks that were recovered to pending after server restart.
-    // These had PRs when the server crashed and need their review loop re-started.
-    // Without this, recovered tasks silently hang in pending forever.
-    //
-    // Each task is re-dispatched in a background tokio task so that permit
-    // acquisition never blocks serve() — if more tasks exist than available
-    // concurrency slots, the background futures will simply wait in queue.
-    {
-        let recovered: Vec<_> = state
-            .core
-            .tasks
-            .list_all()
-            .into_iter()
-            .filter(|t| matches!(t.status, task_runner::TaskStatus::Pending) && t.pr_url.is_some())
-            .collect();
-        if !recovered.is_empty() {
-            tracing::info!(
-                count = recovered.len(),
-                "startup: re-dispatching recovered pending task(s) with PRs"
-            );
-            for task in recovered {
-                let state = state.clone();
-                tokio::spawn(async move {
-                    let pr_url = task.pr_url.as_deref().unwrap_or("");
-                    // Issue 4: robust parsing handles /pull/42/files and #fragment suffixes
-                    let pr_num = match parse_pr_num_from_url(pr_url) {
-                        Some(n) => n,
-                        None => {
-                            // pr_url is present but unparseable (empty, corrupted, or
-                            // non-standard).  Simply returning would leave the task stuck in
-                            // 'pending' forever — fail-close it instead so operators can see
-                            // it and the re-dispatch filter never picks it up again.
-                            tracing::error!(
-                                task_id = ?task.id,
-                                pr_url,
-                                "startup recovery: cannot parse PR number from URL — marking task failed"
-                            );
-                            let bad_url = pr_url.to_owned();
-                            if let Err(e) = task_runner::mutate_and_persist(
-                                &state.core.tasks,
-                                &task.id,
-                                move |s| {
-                                    s.status = task_runner::TaskStatus::Failed;
-                                    s.error = Some(format!(
-                                        "startup recovery: unparseable pr_url: {bad_url}"
-                                    ));
-                                },
-                            )
-                            .await
-                            {
-                                tracing::error!(
-                                    task_id = ?task.id,
-                                    "startup recovery: failed to persist failed status: {e}"
-                                );
-                            }
-                            // Fire completion callback so intake sources (e.g. GitHub Issues
-                            // poller) remove this task from their `dispatched` map. Without
-                            // this the issue stays marked as dispatched forever and will never
-                            // be re-queued, causing a silent production deadlock.
-                            if let Some(cb) = &state.intake.completion_callback {
-                                if let Some(final_state) = state.core.tasks.get(&task.id) {
-                                    cb(final_state).await;
-                                }
-                            }
-                            return;
-                        }
-                    };
-
-                    // Issues 2 & 3: resolve canonical project path from repo name so that
-                    // (a) the correct per-project concurrency bucket is used, and
-                    // (b) req.project is populated so the agent runs in the right worktree.
-                    let project_path = match task.repo.as_deref() {
-                        Some(repo) => {
-                            if let Some(registry) = state.core.project_registry.as_deref() {
-                                match registry.resolve_path(repo).await {
-                                    Ok(Some(p)) => Some(p),
-                                    Ok(None) => Some(std::path::PathBuf::from(repo)),
-                                    Err(e) => {
-                                        tracing::warn!(
-                                            task_id = ?task.id,
-                                            repo,
-                                            "startup recovery: registry lookup failed: {e}, using repo as path"
-                                        );
-                                        Some(std::path::PathBuf::from(repo))
-                                    }
-                                }
-                            } else {
-                                Some(std::path::PathBuf::from(repo))
-                            }
-                        }
-                        None => None,
-                    };
-
-                    let canonical = match task_runner::resolve_canonical_project(project_path).await
-                    {
-                        Ok(c) => c,
-                        Err(e) => {
-                            let reason =
-                                format!("startup recovery: failed to resolve project path: {e}");
-                            tracing::error!(task_id = ?task.id, "{reason}");
-                            if let Err(pe) = task_runner::mutate_and_persist(
-                                &state.core.tasks,
-                                &task.id,
-                                move |s| {
-                                    s.status = task_runner::TaskStatus::Failed;
-                                    s.error = Some(reason);
-                                },
-                            )
-                            .await
-                            {
-                                tracing::error!(
-                                    task_id = ?task.id,
-                                    "startup recovery: failed to persist failed status: {pe}; \
-                                     skipping completion callback to avoid state split"
-                                );
-                                return;
-                            }
-                            if let Some(cb) = &state.intake.completion_callback {
-                                if let Some(final_state) = state.core.tasks.get(&task.id) {
-                                    cb(final_state).await;
-                                }
-                            }
-                            return;
-                        }
-                    };
-                    let project_id = canonical.to_string_lossy().into_owned();
-
-                    // Issue 1: acquire permit here inside the spawned future so serve()
-                    // is never blocked waiting for a concurrency slot.
-                    let permit = match state
-                        .concurrency
-                        .task_queue
-                        .acquire(&project_id, task.priority)
-                        .await
-                    {
-                        Ok(p) => p,
-                        Err(e) => {
-                            let reason = format!(
-                                "startup recovery: failed to acquire concurrency permit: {e}"
-                            );
-                            tracing::error!(task_id = ?task.id, "{reason}");
-                            if let Err(pe) = task_runner::mutate_and_persist(
-                                &state.core.tasks,
-                                &task.id,
-                                move |s| {
-                                    s.status = task_runner::TaskStatus::Failed;
-                                    s.error = Some(reason);
-                                },
-                            )
-                            .await
-                            {
-                                tracing::error!(
-                                    task_id = ?task.id,
-                                    "startup recovery: failed to persist failed status: {pe}; \
-                                     skipping completion callback to avoid state split"
-                                );
-                                return;
-                            }
-                            if let Some(cb) = &state.intake.completion_callback {
-                                if let Some(final_state) = state.core.tasks.get(&task.id) {
-                                    cb(final_state).await;
-                                }
-                            }
-                            return;
-                        }
-                    };
-
-                    let mut req = task_runner::CreateTaskRequest {
-                        pr: Some(pr_num),
-                        project: Some(canonical),
-                        repo: task.repo.clone(),
-                        source: task.source.clone(),
-                        external_id: task.external_id.clone(),
-                        ..Default::default()
-                    };
-                    // Restore persisted execution limits so the recovered task resumes
-                    // with the same budget / timeout guardrails as originally requested.
-                    if let Some(ref settings) = task.request_settings {
-                        settings.apply_to_req(&mut req);
-                    }
-                    // Use the three-tier select_agent() so that an explicit agent
-                    // pin stored in request_settings (Tier 1) and project-level
-                    // defaults (Tier 2a) are honoured, not bypassed by a raw
-                    // complexity dispatch.
-                    let agent = match task_routes::select_agent(
-                        &req,
-                        &state.core.server.agent_registry,
-                        None,
-                    ) {
-                        Ok(a) => a,
-                        Err(e) => {
-                            let reason = format!("startup recovery: failed to select agent: {e}");
-                            tracing::error!(task_id = ?task.id, "{reason}");
-                            if let Err(pe) = task_runner::mutate_and_persist(
-                                &state.core.tasks,
-                                &task.id,
-                                move |s| {
-                                    s.status = task_runner::TaskStatus::Failed;
-                                    s.error = Some(reason);
-                                },
-                            )
-                            .await
-                            {
-                                tracing::error!(
-                                    task_id = ?task.id,
-                                    "startup recovery: failed to persist failed status: {pe}; \
-                                     skipping completion callback to avoid state split"
-                                );
-                                return;
-                            }
-                            if let Some(cb) = &state.intake.completion_callback {
-                                if let Some(final_state) = state.core.tasks.get(&task.id) {
-                                    cb(final_state).await;
-                                }
-                            }
-                            return;
-                        }
-                    };
-                    let (reviewer, _) = resolve_reviewer(
-                        &state.core.server.agent_registry,
-                        &state.core.server.config.agents.review,
-                        agent.name(),
-                    );
-                    state.core.tasks.register_task_stream(&task.id);
-                    task_runner::spawn_preregistered_task(
-                        task.id,
-                        state.core.tasks.clone(),
-                        agent,
-                        reviewer,
-                        Arc::new(state.core.server.config.clone()),
-                        state.engines.skills.clone(),
-                        state.observability.events.clone(),
-                        state.interceptors.clone(),
-                        req,
-                        state.concurrency.workspace_mgr.clone(),
-                        permit,
-                        state.intake.completion_callback.clone(),
-                        None,
-                    )
-                    .await;
-                });
-            }
-        }
-    }
-
-    // Re-dispatch tasks recovered from plan/triage checkpoints but without a PR.
-    // Source A (above) handles tasks with `pr_url` set.  Source B (below) picks up
-    // remaining pending tasks that have a plan or triage checkpoint — these are
-    // issue/prompt tasks interrupted during planning that need to continue execution.
-    // If the DB query fails, log a warning and skip (startup must not abort).
-    {
-        let checkpoint_tasks = match state.core.tasks.pending_tasks_with_checkpoint().await {
-            Ok(pairs) => pairs,
-            Err(e) => {
-                tracing::warn!(
-                    "startup: failed to query checkpoint tasks, \
-                         skipping plan/triage redispatch: {e}"
-                );
-                vec![]
-            }
-        };
-        if !checkpoint_tasks.is_empty() {
-            tracing::info!(
-                count = checkpoint_tasks.len(),
-                "startup: re-dispatching recovered pending task(s) with plan/triage checkpoints"
-            );
-            for (task, _checkpoint) in checkpoint_tasks {
-                let state = state.clone();
-                tokio::spawn(async move {
-                    // Reconstruct request type from the persisted description.
-                    // Description format: "issue #N" → issue task; other → prompt task.
-                    // PR tasks are handled by Source A and never appear here.
-                    let issue_num = task
-                        .description
-                        .as_deref()
-                        .and_then(|d| d.strip_prefix("issue #"))
-                        .and_then(|s| s.split_whitespace().next())
-                        .and_then(|s| s.parse::<u64>().ok());
-
-                    if issue_num.is_none() {
-                        // Prompt tasks store the placeholder "prompt task" in description
-                        // by design — the original prompt text is never persisted for
-                        // privacy.  Re-dispatching with that placeholder would execute
-                        // against the wrong prompt, so mark the task failed instead to
-                        // prevent the same broken recovery on every subsequent restart.
-                        let reason = if task.description.as_deref() == Some("prompt task") {
-                            "prompt task cannot be recovered after restart: \
-                             original prompt text is not persisted"
-                        } else {
-                            "checkpoint task has no parseable issue number — skipping"
-                        };
-                        tracing::warn!(task_id = ?task.id, "{reason}");
-                        let mut failed = task.clone();
-                        failed.status = task_runner::TaskStatus::Failed;
-                        failed.error = Some(reason.to_string());
-                        state.core.tasks.cache.insert(failed.id.clone(), failed);
-                        if let Err(e) = state.core.tasks.persist(&task.id).await {
-                            tracing::warn!(
-                                task_id = ?task.id,
-                                "startup recovery: failed to persist failed status: {e}"
-                            );
-                        }
-                        if let Some(cb) = &state.intake.completion_callback {
-                            if let Some(final_state) = state.core.tasks.get(&task.id) {
-                                cb(final_state).await;
-                            }
-                        }
-                        return;
-                    }
-
-                    let project_path = match task.repo.as_deref() {
-                        Some(repo) => {
-                            if let Some(registry) = state.core.project_registry.as_deref() {
-                                match registry.resolve_path(repo).await {
-                                    Ok(Some(p)) => Some(p),
-                                    Ok(None) => Some(std::path::PathBuf::from(repo)),
-                                    Err(e) => {
-                                        tracing::warn!(
-                                            task_id = ?task.id,
-                                            repo,
-                                            "startup recovery: registry lookup failed: \
-                                             {e}, using repo as path"
-                                        );
-                                        Some(std::path::PathBuf::from(repo))
-                                    }
-                                }
-                            } else {
-                                Some(std::path::PathBuf::from(repo))
-                            }
-                        }
-                        None => None,
-                    };
-
-                    let canonical = match task_runner::resolve_canonical_project(project_path).await
-                    {
-                        Ok(c) => c,
-                        Err(e) => {
-                            let reason =
-                                format!("startup recovery: failed to resolve project path: {e}");
-                            tracing::error!(task_id = ?task.id, "{reason}");
-                            if let Err(pe) = task_runner::mutate_and_persist(
-                                &state.core.tasks,
-                                &task.id,
-                                move |s| {
-                                    s.status = task_runner::TaskStatus::Failed;
-                                    s.error = Some(reason);
-                                },
-                            )
-                            .await
-                            {
-                                tracing::error!(
-                                    task_id = ?task.id,
-                                    "startup recovery: failed to persist failed status: {pe}; \
-                                     skipping completion callback to avoid state split"
-                                );
-                                return;
-                            }
-                            if let Some(cb) = &state.intake.completion_callback {
-                                if let Some(final_state) = state.core.tasks.get(&task.id) {
-                                    cb(final_state).await;
-                                }
-                            }
-                            return;
-                        }
-                    };
-                    let project_id = canonical.to_string_lossy().into_owned();
-
-                    let permit = match state
-                        .concurrency
-                        .task_queue
-                        .acquire(&project_id, task.priority)
-                        .await
-                    {
-                        Ok(p) => p,
-                        Err(e) => {
-                            let reason = format!(
-                                "startup recovery: failed to acquire concurrency permit: {e}"
-                            );
-                            tracing::error!(task_id = ?task.id, "{reason}");
-                            if let Err(pe) = task_runner::mutate_and_persist(
-                                &state.core.tasks,
-                                &task.id,
-                                move |s| {
-                                    s.status = task_runner::TaskStatus::Failed;
-                                    s.error = Some(reason);
-                                },
-                            )
-                            .await
-                            {
-                                tracing::error!(
-                                    task_id = ?task.id,
-                                    "startup recovery: failed to persist failed status: {pe}; \
-                                     skipping completion callback to avoid state split"
-                                );
-                                return;
-                            }
-                            if let Some(cb) = &state.intake.completion_callback {
-                                if let Some(final_state) = state.core.tasks.get(&task.id) {
-                                    cb(final_state).await;
-                                }
-                            }
-                            return;
-                        }
-                    };
-
-                    // issue_num is always Some here — the None branch returned above.
-                    let issue = issue_num.expect("checked above");
-                    let mut req = task_runner::CreateTaskRequest {
-                        issue: Some(issue),
-                        project: Some(canonical),
-                        repo: task.repo.clone(),
-                        source: task.source.clone(),
-                        external_id: task.external_id.clone(),
-                        priority: task.priority,
-                        ..Default::default()
-                    };
-                    // Restore persisted execution limits and any additional prompt
-                    // context so the recovered task resumes with the same settings
-                    // and caller-supplied context as originally requested.
-                    if let Some(ref settings) = task.request_settings {
-                        settings.apply_to_req(&mut req);
-                    }
-
-                    // Use the three-tier select_agent() so that an explicit agent
-                    // pin stored in request_settings (Tier 1) and project-level
-                    // defaults (Tier 2a) are honoured, not bypassed by a raw
-                    // complexity dispatch.
-                    let agent = match task_routes::select_agent(
-                        &req,
-                        &state.core.server.agent_registry,
-                        None,
-                    ) {
-                        Ok(a) => a,
-                        Err(e) => {
-                            let reason = format!("startup recovery: failed to select agent: {e}");
-                            tracing::error!(task_id = ?task.id, "{reason}");
-                            if let Err(pe) = task_runner::mutate_and_persist(
-                                &state.core.tasks,
-                                &task.id,
-                                move |s| {
-                                    s.status = task_runner::TaskStatus::Failed;
-                                    s.error = Some(reason);
-                                },
-                            )
-                            .await
-                            {
-                                tracing::error!(
-                                    task_id = ?task.id,
-                                    "startup recovery: failed to persist failed status: {pe}; \
-                                     skipping completion callback to avoid state split"
-                                );
-                                return;
-                            }
-                            if let Some(cb) = &state.intake.completion_callback {
-                                if let Some(final_state) = state.core.tasks.get(&task.id) {
-                                    cb(final_state).await;
-                                }
-                            }
-                            return;
-                        }
-                    };
-                    let (reviewer, _) = resolve_reviewer(
-                        &state.core.server.agent_registry,
-                        &state.core.server.config.agents.review,
-                        agent.name(),
-                    );
-                    state.core.tasks.register_task_stream(&task.id);
-                    task_runner::spawn_preregistered_task(
-                        task.id,
-                        state.core.tasks.clone(),
-                        agent,
-                        reviewer,
-                        Arc::new(state.core.server.config.clone()),
-                        state.engines.skills.clone(),
-                        state.observability.events.clone(),
-                        state.interceptors.clone(),
-                        req,
-                        state.concurrency.workspace_mgr.clone(),
-                        permit,
-                        state.intake.completion_callback.clone(),
-                        None,
-                    )
-                    .await;
-                });
-            }
-        }
-    }
-
-    let initial_grade = {
-        let events = state
-            .observability
-            .events
-            .query(&harness_core::types::EventFilters::default())
-            .await
-            .unwrap_or_default();
-        // Use violations from the most recent scan (identified by the latest rule_scan session_id)
-        // rather than all historical rule_check events, to avoid permanently depressing the grade.
-        let violation_count = events
-            .iter()
-            .rev()
-            .find(|e| e.hook == "rule_scan")
-            .map(|scan| {
-                events
-                    .iter()
-                    .filter(|e| e.hook == "rule_check" && e.session_id == scan.session_id)
-                    .count()
-            })
-            .unwrap_or(0);
-        harness_observe::quality::QualityGrader::grade(&events, violation_count).grade
-    };
-    crate::scheduler::Scheduler::from_grade(initial_grade).start(state.clone());
-    // Pass the pre-built GitHub pollers from AppState to the orchestrator so
-    // both share the same Arc instances and on_task_complete operates on the
-    // live poller's dispatched map.
-    let github_sources = state.intake.github_pollers.clone();
-    crate::intake::build_orchestrator(
-        &state.core.server.config.intake,
-        Some(&expand_tilde(&state.core.server.config.server.data_dir)),
-        state.intake.feishu_intake.clone(),
-        github_sources,
-    )
-    .start(state.clone());
-
-    let app = Router::new()
-        .route("/", get(crate::dashboard::index))
-        .route("/favicon.ico", get(crate::dashboard::favicon))
-        .route("/health", get(health_check))
-        .route("/rpc", post(handle_rpc))
-        .route("/ws", get(crate::websocket::ws_handler))
-        .route("/tasks", post(task_routes::create_task))
-        .route("/tasks", get(list_tasks))
-        .route("/tasks/batch", post(task_routes::create_tasks_batch))
-        .route("/tasks/{id}", get(get_task))
-        .route("/tasks/{id}/cancel", post(task_routes::cancel_task))
-        .route("/tasks/{id}/artifacts", get(get_task_artifacts))
-        .route("/tasks/{id}/stream", get(stream_task_sse))
-        .route(
-            "/projects",
-            post(crate::handlers::projects::register_project)
-                .get(crate::handlers::projects::list_projects),
-        )
-        .route(
-            "/projects/{id}",
-            get(crate::handlers::projects::get_project)
-                .delete(crate::handlers::projects::delete_project),
-        )
-        .route("/projects/queue-stats", get(project_queue_stats))
-        .route("/api/dashboard", get(crate::handlers::dashboard::dashboard))
-        .route("/api/intake", get(intake_status))
-        .route(
-            "/api/runtime-hosts",
-            get(crate::handlers::runtime_hosts::list_runtime_hosts),
-        )
-        .route(
-            "/api/runtime-hosts/register",
-            post(crate::handlers::runtime_hosts::register_runtime_host),
-        )
-        .route(
-            "/api/runtime-hosts/{host_id}/heartbeat",
-            post(crate::handlers::runtime_hosts::heartbeat_runtime_host),
-        )
-        .route(
-            "/api/runtime-hosts/{host_id}/deregister",
-            post(crate::handlers::runtime_hosts::deregister_runtime_host),
-        )
-        .route(
-            "/api/runtime-hosts/{host_id}/tasks/claim",
-            post(crate::handlers::runtime_hosts::claim_task_for_runtime_host),
-        )
-        .route(
-            "/api/runtime-hosts/{host_id}/projects",
-            get(crate::handlers::runtime_project_cache::list_runtime_host_projects),
-        )
-        .route(
-            "/api/runtime-hosts/{host_id}/projects/sync",
-            post(crate::handlers::runtime_project_cache::sync_runtime_host_projects),
-        )
-        .route(
-            "/api/token-usage",
-            get(crate::handlers::token_usage::token_usage),
-        )
-        .route(
-            "/webhook",
-            post(github_webhook).layer(DefaultBodyLimit::max(
-                state.core.server.config.server.max_webhook_body_bytes,
-            )),
-        )
-        .route(
-            "/webhook/feishu",
-            post(crate::intake::feishu::feishu_webhook).layer(DefaultBodyLimit::max(
-                state.core.server.config.server.max_webhook_body_bytes,
-            )),
-        )
-        .route(
-            "/signals",
-            post(ingest_signal).layer(DefaultBodyLimit::max(
-                state.core.server.config.server.max_webhook_body_bytes,
-            )),
-        )
-        .route("/auth/reset-password", post(password_reset))
-        .layer(middleware::from_fn_with_state(
-            state.clone(),
-            auth::api_auth_middleware,
-        ))
-        .with_state(state.clone());
-
-    let listener = tokio::net::TcpListener::bind(addr).await?;
-    let ws_shutdown_tx = state.notifications.ws_shutdown_tx.clone();
-    let serve_result = axum::serve(listener, app)
-        .with_graceful_shutdown(async move {
-            shutdown_signal().await;
-            tracing::info!("server shutting down: closing WebSocket connections");
-            ws_shutdown_tx.send(()).ok();
-        })
-        .await;
-    tracing::info!("server shutting down");
-    state.observability.events.shutdown().await;
-    serve_result?;
-    Ok(())
-}
-
-async fn shutdown_signal() {
-    let ctrl_c = async {
-        tokio::signal::ctrl_c()
-            .await
-            .expect("failed to install Ctrl+C handler");
-    };
-
-    #[cfg(unix)]
-    let terminate = async {
-        tokio::signal::unix::signal(tokio::signal::unix::SignalKind::terminate())
-            .expect("failed to install SIGTERM handler")
-            .recv()
-            .await;
-    };
-
-    #[cfg(not(unix))]
-    let terminate = std::future::pending::<()>();
-
-    tokio::select! {
-        _ = ctrl_c => tracing::info!("received Ctrl+C"),
-        _ = terminate => tracing::info!("received SIGTERM"),
-    }
-}
-
-async fn health_check(State(state): State<Arc<AppState>>) -> Json<serde_json::Value> {
-    let count = state.core.tasks.count();
-    Json(json!({"status": "ok", "tasks": count}))
-}
-
-/// GET /projects — per-project queue stats alongside the global queue summary.
-async fn project_queue_stats(State(state): State<Arc<AppState>>) -> Json<serde_json::Value> {
-    let tq = &state.concurrency.task_queue;
-    let projects: serde_json::Map<String, serde_json::Value> = tq
-        .all_project_stats()
-        .into_iter()
-        .map(|(id, s)| {
-            (
-                id,
-                json!({
-                    "running": s.running,
-                    "queued": s.queued,
-                    "limit": s.limit,
-                }),
-            )
-        })
-        .collect();
-    Json(json!({
-        "global": {
-            "running": tq.running_count(),
-            "queued": tq.queued_count(),
-            "limit": tq.global_limit(),
-        },
-        "projects": projects,
-    }))
-}
-
-async fn handle_rpc(State(state): State<Arc<AppState>>, Json(req): Json<RpcRequest>) -> Response {
-    match router::handle_request(&state, req).await {
-        Some(resp) => (StatusCode::OK, Json(resp)).into_response(),
-        None => StatusCode::NO_CONTENT.into_response(),
-    }
-}
-
-async fn github_webhook(
-    State(state): State<Arc<AppState>>,
-    headers: HeaderMap,
-    body: Bytes,
-) -> (StatusCode, Json<serde_json::Value>) {
-    let secret = match state
-        .core
-        .server
-        .config
-        .server
-        .github_webhook_secret
-        .as_deref()
-    {
-        Some("") => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({"error": "invalid server.github_webhook_secret configuration"})),
-            )
-        }
-        Some(secret) => secret,
-        None => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({"error": "missing server.github_webhook_secret configuration"})),
-            )
-        }
-    };
-    let signature = match headers
-        .get("x-hub-signature-256")
-        .and_then(|value| value.to_str().ok())
-    {
-        Some(signature) => signature,
-        None => {
-            return (
-                StatusCode::UNAUTHORIZED,
-                Json(json!({"error": "missing header x-hub-signature-256"})),
-            )
-        }
-    };
-    if !crate::webhook::verify_github_signature(secret, signature, body.as_ref()) {
-        return (
-            StatusCode::UNAUTHORIZED,
-            Json(json!({"error": "invalid webhook signature"})),
-        );
-    }
-
-    let event = match headers
-        .get("x-github-event")
-        .and_then(|value| value.to_str().ok())
-    {
-        Some(event) => event,
-        None => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(json!({"error": "missing header x-github-event"})),
-            )
-        }
-    };
-    if !crate::webhook::is_valid_github_event_name(event) {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(json!({"error": "invalid header x-github-event"})),
-        );
-    }
-
-    let (request, reason) =
-        match crate::webhook::parse_github_webhook_task_request(event, body.as_ref()) {
-            Ok(parsed) => parsed,
-            Err(error) => return (StatusCode::BAD_REQUEST, Json(json!({ "error": error }))),
-        };
-
-    let Some(mut req) = request else {
-        return (
-            StatusCode::OK,
-            Json(json!({
-                "status": "ignored",
-                "reason": reason,
-            })),
-        );
-    };
-
-    if req.project.is_none() {
-        req.project = Some(state.core.project_root.clone());
-    }
-
-    match task_routes::enqueue_task(&state, req).await {
-        Ok(task_id) => (
-            StatusCode::ACCEPTED,
-            Json(json!({
-                "status": "accepted",
-                "reason": reason,
-                "task_id": task_id.0,
-            })),
-        ),
-        Err(crate::services::execution::EnqueueTaskError::BadRequest(error)) => {
-            (StatusCode::BAD_REQUEST, Json(json!({ "error": error })))
-        }
-        Err(crate::services::execution::EnqueueTaskError::Internal(error)) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(json!({ "error": error })),
-        ),
-    }
-}
-
-async fn list_tasks(State(state): State<Arc<AppState>>) -> Response {
-    match state.core.tasks.list_all_summaries_with_terminal().await {
-        Ok(summaries) => Json(summaries).into_response(),
-        Err(e) => {
-            tracing::error!("list_tasks: database error: {e}");
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({"error": "internal server error"})),
-            )
-                .into_response()
-        }
-    }
-}
-
-async fn get_task(State(state): State<Arc<AppState>>, Path(id): Path<String>) -> Response {
-    match state
-        .core
-        .tasks
-        .get_with_db_fallback(&harness_core::types::TaskId(id))
-        .await
-    {
-        Ok(Some(task)) => Json(task).into_response(),
-        Ok(None) => (
-            StatusCode::NOT_FOUND,
-            Json(json!({"error": "task not found"})),
-        )
-            .into_response(),
-        Err(e) => {
-            tracing::error!("get_task: database error: {e}");
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({"error": "internal server error"})),
-            )
-                .into_response()
-        }
-    }
-}
-
-/// GET /tasks/{id}/artifacts — all persisted artifacts for a task.
-async fn get_task_artifacts(
-    State(state): State<Arc<AppState>>,
-    Path(id): Path<String>,
-) -> Response {
-    let task_id = harness_core::types::TaskId(id);
-    match state.core.tasks.get_with_db_fallback(&task_id).await {
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(json!({"error": "task not found"})),
-            )
-                .into_response();
-        }
-        Err(e) => {
-            tracing::error!("get_task_artifacts: database error: {e}");
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({"error": "internal server error"})),
-            )
-                .into_response();
-        }
-        Ok(Some(_)) => {}
-    }
-    match state.core.tasks.list_artifacts(&task_id).await {
-        Ok(artifacts) => Json(artifacts).into_response(),
-        Err(e) => {
-            tracing::error!("get_task_artifacts: list artifacts error: {e}");
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({"error": "internal server error"})),
-            )
-                .into_response()
-        }
-    }
-}
-
-/// GET /tasks/{id}/stream — real-time SSE stream of agent execution events.
-///
-/// Subscribes to the task's broadcast channel and forwards each [`StreamItem`]
-/// as a JSON-encoded SSE data event. The stream ends when the task completes
-/// (channel closed). If the receiver lags, a synthetic "lag" event is emitted
-/// noting how many events were dropped; streaming then continues.
-async fn stream_task_sse(State(state): State<Arc<AppState>>, Path(id): Path<String>) -> Response {
-    let task_id = harness_core::types::TaskId(id);
-
-    let rx = match state.core.tasks.subscribe_task_stream(&task_id) {
-        Some(rx) => rx,
-        None => {
-            match state.core.tasks.get_with_db_fallback(&task_id).await {
-                Ok(None) => {
-                    return (
-                        StatusCode::NOT_FOUND,
-                        Json(json!({"error": "task not found"})),
-                    )
-                        .into_response();
-                }
-                Err(e) => {
-                    tracing::error!("stream_task_sse: database error: {e}");
-                    return (
-                        StatusCode::INTERNAL_SERVER_ERROR,
-                        Json(json!({"error": "internal server error"})),
-                    )
-                        .into_response();
-                }
-                Ok(Some(_)) => {}
-            }
-            // Task exists but stream already closed (task completed before client connected).
-            let stream = futures::stream::empty::<Result<Event, std::convert::Infallible>>();
-            return Sse::new(stream).into_response();
-        }
-    };
-
-    let stream = futures::stream::unfold(rx, |mut rx| async move {
-        match rx.recv().await {
-            Ok(item) => {
-                let data = match serde_json::to_string(&item) {
-                    Ok(s) => s,
-                    Err(e) => {
-                        tracing::warn!("sse: failed to serialize event: {e}");
-                        String::new()
-                    }
-                };
-                Some((
-                    Ok::<Event, std::convert::Infallible>(Event::default().data(data)),
-                    rx,
-                ))
-            }
-            Err(tokio::sync::broadcast::error::RecvError::Closed) => None,
-            Err(tokio::sync::broadcast::error::RecvError::Lagged(n)) => {
-                let event = Event::default()
-                    .event("lag")
-                    .data(format!("dropped {n} events due to slow consumer"));
-                Some((Ok(event), rx))
-            }
-        }
-    });
-
-    // Send a heartbeat comment every 30 s so reverse proxies (nginx default
-    // 60 s idle timeout) don't drop the connection while the agent is silent.
-    Sse::new(stream)
-        .keep_alive(
-            KeepAlive::new()
-                .interval(std::time::Duration::from_secs(30))
-                .text("heartbeat"),
-        )
-        .into_response()
-}
-
-/// GET /api/intake — current status of all intake channels and recent dispatches.
-async fn intake_status(State(state): State<Arc<AppState>>) -> Json<serde_json::Value> {
-    let intake_config = &state.core.server.config.intake;
-    let all_tasks = state.core.tasks.list_all();
-
-    let github_active: u64 = all_tasks
-        .iter()
-        .filter(|t| {
-            t.source.as_deref() == Some("github")
-                && !matches!(
-                    t.status,
-                    task_runner::TaskStatus::Done | task_runner::TaskStatus::Failed
-                )
-        })
-        .count() as u64;
-
-    let feishu_active: u64 = all_tasks
-        .iter()
-        .filter(|t| {
-            t.source.as_deref() == Some("feishu")
-                && !matches!(
-                    t.status,
-                    task_runner::TaskStatus::Done | task_runner::TaskStatus::Failed
-                )
-        })
-        .count() as u64;
-
-    let dashboard_active: u64 = all_tasks
-        .iter()
-        .filter(|t| {
-            (t.source.as_deref() == Some("dashboard") || t.source.is_none())
-                && !matches!(
-                    t.status,
-                    task_runner::TaskStatus::Done | task_runner::TaskStatus::Failed
-                )
-        })
-        .count() as u64;
-
-    let github_channel = json!({
-        "name": "github",
-        "enabled": intake_config.github.as_ref().map(|c| c.enabled).unwrap_or(false),
-        "repo": intake_config.github.as_ref().map(|c| c.repo.as_str()).unwrap_or(""),
-        "active": github_active,
-    });
-
-    let feishu_channel = json!({
-        "name": "feishu",
-        "enabled": state.intake.feishu_intake.is_some(),
-        "keyword": intake_config.feishu.as_ref().map(|c| c.trigger_keyword.as_str()).unwrap_or(""),
-        "active": feishu_active,
-    });
-
-    let dashboard_channel = json!({
-        "name": "dashboard",
-        "enabled": true,
-        "active": dashboard_active,
-    });
-
-    let mut recent_dispatches: Vec<serde_json::Value> = all_tasks
-        .iter()
-        .filter(|t| t.source.is_some())
-        .map(|t| {
-            json!({
-                "source": t.source,
-                "external_id": t.external_id,
-                "task_id": t.id.0,
-                "status": serde_json::to_value(&t.status).unwrap_or(json!("unknown")),
-                "pr_url": t.pr_url,
-            })
-        })
-        .collect();
-    recent_dispatches.truncate(10);
-
-    Json(json!({
-        "channels": [github_channel, feishu_channel, dashboard_channel],
-        "recent_dispatches": recent_dispatches,
-    }))
-}
-
-/// Request body for `POST /signals`.
-#[derive(serde::Deserialize)]
-struct IngestSignalRequest {
-    source: String,
-    #[serde(default)]
-    severity: Option<harness_core::types::Severity>,
-    payload: serde_json::Value,
-}
-
-/// Infer severity from a GitHub webhook payload: CI failure → High, changes_requested → Medium.
-fn infer_github_severity(payload: &serde_json::Value) -> Option<harness_core::types::Severity> {
-    if let Some(obj) = payload.as_object() {
-        // check_run completed with failure
-        if let (Some(action), Some(check_run)) = (
-            obj.get("action").and_then(|v| v.as_str()),
-            obj.get("check_run"),
-        ) {
-            if action == "completed" {
-                let conclusion = check_run
-                    .get("conclusion")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("");
-                if conclusion == "failure" {
-                    return Some(harness_core::types::Severity::High);
-                }
-            }
-        }
-        // pull_request_review with changes_requested
-        if let Some(review) = obj.get("review") {
-            let state = review.get("state").and_then(|v| v.as_str()).unwrap_or("");
-            if state.eq_ignore_ascii_case("changes_requested") {
-                return Some(harness_core::types::Severity::Medium);
-            }
-        }
-    }
-    None
-}
-
-/// POST /signals — ingest an external signal (CI failure, review feedback, etc.).
-///
-/// Validates the `x-hub-signature-256` HMAC-SHA256 header using the configured
-/// `server.github_webhook_secret`. Rate-limited to 100 requests per source per minute.
-async fn ingest_signal(
-    State(state): State<Arc<AppState>>,
-    headers: HeaderMap,
-    body: Bytes,
-) -> (StatusCode, Json<serde_json::Value>) {
-    let secret = match state
-        .core
-        .server
-        .config
-        .server
-        .github_webhook_secret
-        .as_deref()
-    {
-        Some("") | None => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(json!({"error": "server.github_webhook_secret not configured"})),
-            )
-        }
-        Some(s) => s,
-    };
-
-    let signature = match headers
-        .get("x-hub-signature-256")
-        .and_then(|v| v.to_str().ok())
-    {
-        Some(sig) => sig,
-        None => {
-            return (
-                StatusCode::UNAUTHORIZED,
-                Json(json!({"error": "missing header x-hub-signature-256"})),
-            )
-        }
-    };
-
-    if !crate::webhook::verify_github_signature(secret, signature, body.as_ref()) {
-        return (
-            StatusCode::UNAUTHORIZED,
-            Json(json!({"error": "invalid webhook signature"})),
-        );
-    }
-
-    let req: IngestSignalRequest = match serde_json::from_slice(&body) {
-        Ok(r) => r,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(json!({"error": format!("invalid payload: {e}")})),
-            )
-        }
-    };
-
-    if !state
-        .observability
-        .signal_rate_limiter
-        .check_and_increment(&req.source)
-    {
-        return (
-            StatusCode::TOO_MANY_REQUESTS,
-            Json(json!({"error": "rate limit exceeded: max 100 signals per minute per source"})),
-        );
-    }
-
-    let severity = req.severity.unwrap_or_else(|| {
-        if req.source == "github" {
-            infer_github_severity(&req.payload).unwrap_or(harness_core::types::Severity::Low)
-        } else {
-            harness_core::types::Severity::Low
-        }
-    });
-
-    let signal =
-        harness_core::types::ExternalSignal::new(req.source.clone(), severity, req.payload.clone());
-    let signal_id = signal.id.clone();
-
-    if let Err(e) = state.observability.events.log_external_signal(&signal) {
-        tracing::error!(source = %req.source, "failed to store external signal: {e}");
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(json!({"error": "failed to store signal"})),
-        );
-    }
-
-    tracing::info!(
-        source = %req.source,
-        severity = ?severity,
-        signal_id = %signal_id,
-        "external signal ingested"
-    );
-
-    (
-        StatusCode::OK,
-        Json(json!({"status": "accepted", "id": signal_id.as_str()})),
-    )
-}
-
-#[derive(serde::Deserialize)]
-struct PasswordResetRequest {
-    email: String,
-}
-
-/// POST /auth/reset-password — initiate a password reset.
-///
-/// Rate-limited per email address to prevent enumeration and brute-force.
-/// Always returns a generic success response regardless of whether the email
-/// exists, to avoid leaking account information.
-async fn password_reset(
-    State(state): State<Arc<AppState>>,
-    Json(req): Json<PasswordResetRequest>,
-) -> (StatusCode, Json<serde_json::Value>) {
-    let email = req.email.trim().to_lowercase();
-    if email.is_empty() {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(json!({"error": "email is required"})),
-        );
-    }
-
-    let limit = state
-        .core
-        .server
-        .config
-        .server
-        .password_reset_rate_limit_per_hour;
-
-    if !state
-        .observability
-        .password_reset_rate_limiter
-        .check_and_increment(&email)
-    {
-        return (
-            StatusCode::TOO_MANY_REQUESTS,
-            Json(json!({
-                "error": format!(
-                    "rate limit exceeded: max {} password reset requests per hour",
-                    limit
-                )
-            })),
-        );
-    }
-
-    tracing::info!(
-        email_hash = %format!("{:x}", {
-            use std::hash::{Hash, Hasher};
-            let mut h = std::collections::hash_map::DefaultHasher::new();
-            email.hash(&mut h);
-            h.finish()
-        }),
-        "password reset requested"
-    );
-
-    (
-        StatusCode::OK,
-        Json(
-            json!({"status": "ok", "message": "If that email is registered, a reset link has been sent."}),
-        ),
-    )
-}
+pub(crate) mod types;
+
+pub(crate) use helpers::{parse_pr_num_from_url, resolve_reviewer};
+pub use init::build_app_state;
+pub(crate) use init::build_completion_callback;
+pub use server::serve;
+pub use types::{
+    AppState, ConcurrencyServices, CoreServices, EngineServices, IntakeServices,
+    NotificationServices, ObservabilityServices,
+};
 
 #[cfg(test)]
-mod startup_tests {
-    use super::build_app_state;
-    use crate::{
-        server::HarnessServer,
-        test_helpers::{HomeGuard, HOME_LOCK},
-        thread_manager::ThreadManager,
-    };
-    use harness_agents::registry::AgentRegistry;
-    use harness_core::{
-        config::HarnessConfig, types::EventFilters, types::RuleId, types::Severity,
-        types::SkillLocation, types::Violation,
-    };
-    use std::sync::Arc;
-
-    #[tokio::test]
-    async fn persisted_skills_survive_restart() -> anyhow::Result<()> {
-        // Hold the shared HOME_LOCK so no sibling test races on HOME.
-        let _lock = HOME_LOCK.lock().await;
-
-        let sandbox = tempfile::tempdir()?;
-        let project_root = sandbox.path().join("project");
-        std::fs::create_dir_all(&project_root)?;
-        let data_dir = sandbox.path().join("data");
-
-        // Redirect HOME to an empty sandbox directory so that
-        // $HOME/.harness/skills/ cannot shadow the persisted skill under
-        // data_dir, keeping the test isolated from machine state.
-        let fake_home = sandbox.path().join("home");
-        std::fs::create_dir_all(&fake_home)?;
-        // SAFETY: HOME_LOCK is held above; HomeGuard::drop restores HOME
-        // unconditionally, even when an assertion below panics.
-        let _env_guard = unsafe { HomeGuard::set(&fake_home) };
-
-        let startup = |project_root: &std::path::Path, data_dir: &std::path::Path| {
-            let project_root = project_root.to_path_buf();
-            let data_dir = data_dir.to_path_buf();
-            async move {
-                let mut config = HarnessConfig::default();
-                config.server.project_root = project_root;
-                config.server.data_dir = data_dir;
-                let server = Arc::new(HarnessServer::new(
-                    config,
-                    ThreadManager::new(),
-                    AgentRegistry::new("test"),
-                ));
-                build_app_state(server).await
-            }
-        };
-
-        // First startup: create a skill so it gets persisted to disk.
-        {
-            let state = startup(&project_root, &data_dir).await?;
-            let mut skills = state.engines.skills.write().await;
-            skills.create("my-test-skill".to_string(), "# My test skill".to_string());
-        }
-
-        // Assert the skill file was physically written to data_dir/skills/
-        // before the second startup, catching a broken persist_dir path early.
-        let persisted_path = data_dir.join("skills").join("my-test-skill.md");
-        assert!(
-            persisted_path.exists(),
-            "expected skill file to be written to {}",
-            persisted_path.display()
-        );
-
-        // Second startup: verify the persisted skill is reloaded via discover().
-        {
-            let state = startup(&project_root, &data_dir).await?;
-            let skills = state.engines.skills.read().await;
-            let reloaded = skills
-                .list()
-                .iter()
-                .find(|s| s.name == "my-test-skill")
-                .ok_or_else(|| {
-                    anyhow::anyhow!("expected persisted skill to be reloaded after restart")
-                })?;
-            // Skills persisted via store.create() are stored in data_dir/skills/
-            // and loaded with SkillLocation::User so they can override same-named
-            // builtins after restart (User priority > System priority).
-            assert_eq!(
-                reloaded.location,
-                SkillLocation::User,
-                "reloaded skill has location {:?}; expected User (data_dir/skills/)",
-                reloaded.location
-            );
-        }
-
-        Ok(())
-        // _env_guard dropped here → HOME restored unconditionally
-        // _lock dropped here → next test may proceed
-    }
-
-    #[tokio::test]
-    async fn build_app_state_auto_registers_builtin_guard() -> anyhow::Result<()> {
-        let _lock = HOME_LOCK.lock().await;
-        let sandbox = tempfile::tempdir()?;
-        let project_root = sandbox.path().join("project");
-        std::fs::create_dir_all(&project_root)?;
-
-        let mut config = HarnessConfig::default();
-        config.server.project_root = project_root;
-        config.server.data_dir = sandbox.path().join("data");
-
-        let server = Arc::new(HarnessServer::new(
-            config,
-            ThreadManager::new(),
-            AgentRegistry::new("test"),
-        ));
-        let state = build_app_state(server).await?;
-        let rules = state.engines.rules.read().await;
-
-        assert!(
-            rules
-                .guards()
-                .iter()
-                .any(|guard| guard.id.as_str() == harness_rules::engine::BUILTIN_BASELINE_GUARD_ID),
-            "expected build_app_state to auto-register builtin guard"
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn build_app_state_registers_startup_project_metadata() -> anyhow::Result<()> {
-        let _lock = HOME_LOCK.lock().await;
-        let sandbox = tempfile::tempdir()?;
-        let project_root = sandbox.path().join("project");
-        std::fs::create_dir_all(&project_root)?;
-
-        let mut config = HarnessConfig::default();
-        config.server.project_root = project_root.clone();
-        config.server.data_dir = sandbox.path().join("data");
-
-        let mut server =
-            HarnessServer::new(config, ThreadManager::new(), AgentRegistry::new("test"));
-        server.startup_projects = vec![harness_core::config::ProjectEntry {
-            name: "named".to_string(),
-            root: project_root.clone(),
-            default: false,
-            default_agent: Some("codex".to_string()),
-            max_concurrent: Some(7),
-        }];
-
-        let state = build_app_state(Arc::new(server)).await?;
-        let registry = state
-            .core
-            .project_registry
-            .as_ref()
-            .expect("project registry should be initialized");
-        let project = registry
-            .get("named")
-            .await?
-            .expect("startup project should be registered");
-
-        assert_eq!(project.default_agent.as_deref(), Some("codex"));
-        assert_eq!(project.max_concurrent, Some(7));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn build_app_state_registers_default_project_metadata_from_startup_entry(
-    ) -> anyhow::Result<()> {
-        let _lock = HOME_LOCK.lock().await;
-        let sandbox = tempfile::tempdir()?;
-        let project_root = sandbox.path().join("project");
-        std::fs::create_dir_all(&project_root)?;
-
-        let mut config = HarnessConfig::default();
-        config.server.project_root = project_root.clone();
-        config.server.data_dir = sandbox.path().join("data");
-
-        let mut server =
-            HarnessServer::new(config, ThreadManager::new(), AgentRegistry::new("test"));
-        server.startup_projects = vec![harness_core::config::ProjectEntry {
-            name: "configured-default".to_string(),
-            root: project_root.clone(),
-            default: true,
-            default_agent: Some("claude".to_string()),
-            max_concurrent: Some(3),
-        }];
-
-        let state = build_app_state(Arc::new(server)).await?;
-        let registry = state
-            .core
-            .project_registry
-            .as_ref()
-            .expect("project registry should be initialized");
-        let project = registry
-            .get("default")
-            .await?
-            .expect("default project should be registered");
-
-        assert_eq!(project.root.canonicalize()?, project_root.canonicalize()?);
-        assert_eq!(project.default_agent.as_deref(), Some("claude"));
-        assert_eq!(project.max_concurrent, Some(3));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn build_app_state_preserves_partial_startup_project_metadata() -> anyhow::Result<()> {
-        let _lock = HOME_LOCK.lock().await;
-        let sandbox = tempfile::tempdir()?;
-        let project_root = sandbox.path().join("project");
-        std::fs::create_dir_all(&project_root)?;
-
-        let mut config = HarnessConfig::default();
-        config.server.project_root = project_root.clone();
-        config.server.data_dir = sandbox.path().join("data");
-
-        let mut server =
-            HarnessServer::new(config, ThreadManager::new(), AgentRegistry::new("test"));
-        server.startup_projects = vec![harness_core::config::ProjectEntry {
-            name: "partial".to_string(),
-            root: project_root,
-            default: false,
-            default_agent: Some("claude".to_string()),
-            max_concurrent: None,
-        }];
-
-        let state = build_app_state(Arc::new(server)).await?;
-        let registry = state
-            .core
-            .project_registry
-            .as_ref()
-            .expect("project registry should be initialized");
-        let project = registry
-            .get("partial")
-            .await?
-            .expect("startup project should be registered");
-
-        assert_eq!(project.default_agent.as_deref(), Some("claude"));
-        assert_eq!(project.max_concurrent, None);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn build_app_state_ignores_stale_default_project_metadata_for_different_root(
-    ) -> anyhow::Result<()> {
-        let _lock = HOME_LOCK.lock().await;
-        let sandbox = tempfile::tempdir()?;
-        let startup_root = sandbox.path().join("startup-project");
-        let override_root = sandbox.path().join("override-project");
-        std::fs::create_dir_all(&startup_root)?;
-        std::fs::create_dir_all(&override_root)?;
-
-        let mut config = HarnessConfig::default();
-        config.server.project_root = override_root.clone();
-        config.server.data_dir = sandbox.path().join("data");
-
-        let mut server =
-            HarnessServer::new(config, ThreadManager::new(), AgentRegistry::new("test"));
-        let startup_default_project = harness_core::config::ProjectEntry {
-            name: "configured-default".to_string(),
-            root: startup_root.clone(),
-            default: true,
-            default_agent: Some("claude".to_string()),
-            max_concurrent: Some(3),
-        };
-        server.startup_projects = vec![startup_default_project.clone()];
-        server.startup_default_project = Some(startup_default_project);
-
-        let state = build_app_state(Arc::new(server)).await?;
-        let registry = state
-            .core
-            .project_registry
-            .as_ref()
-            .expect("project registry should be initialized");
-        let project = registry
-            .get("default")
-            .await?
-            .expect("default project should be registered");
-
-        assert_eq!(project.root.canonicalize()?, override_root.canonicalize()?);
-        assert_eq!(project.default_agent, None);
-        assert_eq!(project.max_concurrent, None);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn startup_grade_uses_latest_rule_scan_session_for_violation_count() -> anyhow::Result<()>
-    {
-        let _lock = HOME_LOCK.lock().await;
-        let sandbox = tempfile::tempdir()?;
-        let project_root = sandbox.path().join("project");
-        std::fs::create_dir_all(&project_root)?;
-        let data_dir = sandbox.path().join("data");
-
-        // Redirect HOME so build_app_state does not read from the real user home.
-        let fake_home = sandbox.path().join("home");
-        std::fs::create_dir_all(&fake_home)?;
-        // SAFETY: HOME_LOCK is held above; HomeGuard::drop restores HOME unconditionally.
-        let _env_guard = unsafe { HomeGuard::set(&fake_home) };
-
-        let mut config = HarnessConfig::default();
-        config.server.project_root = project_root.clone();
-        config.server.data_dir = data_dir;
-        let server = Arc::new(HarnessServer::new(
-            config,
-            ThreadManager::new(),
-            AgentRegistry::new("test"),
-        ));
-        let state = build_app_state(server).await?;
-
-        // First scan: persist 5 violations (old session — must NOT count at startup).
-        let old_violations: Vec<Violation> = (0..5)
-            .map(|i| Violation {
-                rule_id: RuleId::from_str(&format!("U-{i:02}")),
-                file: std::path::PathBuf::from("src/old.rs"),
-                line: Some(i + 1),
-                message: format!("old violation {i}"),
-                severity: Severity::Low,
-            })
-            .collect();
-        state
-            .observability
-            .events
-            .persist_rule_scan(&project_root, &old_violations)
-            .await;
-
-        // Second scan: persist 2 violations (latest session — must be used for startup grade).
-        let new_violations = vec![
-            Violation {
-                rule_id: RuleId::from_str("SEC-01"),
-                file: std::path::PathBuf::from("src/lib.rs"),
-                line: Some(1),
-                message: "new critical violation".to_string(),
-                severity: Severity::Critical,
-            },
-            Violation {
-                rule_id: RuleId::from_str("SEC-02"),
-                file: std::path::PathBuf::from("src/main.rs"),
-                line: None,
-                message: "another new violation".to_string(),
-                severity: Severity::High,
-            },
-        ];
-        state
-            .observability
-            .events
-            .persist_rule_scan(&project_root, &new_violations)
-            .await;
-
-        // Replicate the exact startup grade logic from serve() (lines 687-697).
-        let events = state
-            .observability
-            .events
-            .query(&EventFilters::default())
-            .await
-            .unwrap_or_default();
-        let violation_count = events
-            .iter()
-            .rev()
-            .find(|e| e.hook == "rule_scan")
-            .map(|scan| {
-                events
-                    .iter()
-                    .filter(|e| e.hook == "rule_check" && e.session_id == scan.session_id)
-                    .count()
-            })
-            .unwrap_or(0);
-
-        // Must count only the latest scan session (2 violations), not historical total (7).
-        assert_eq!(
-            violation_count,
-            new_violations.len(),
-            "startup grade must use latest scan session ({} violations), not historical total ({})",
-            new_violations.len(),
-            old_violations.len() + new_violations.len(),
-        );
-
-        Ok(())
-        // _env_guard dropped here → HOME restored unconditionally
-        // _lock dropped here → next test may proceed
-    }
-}
+mod startup_tests;
 
 #[cfg(test)]
 mod tests;
