+++ crates/harness-server/src/http/handlers.rs
@@ -1,0 +1,315 @@
+use super::types::AppState;
+use axum::{
+    body::Bytes,
+    extract::{Path, State},
+    http::{HeaderMap, StatusCode},
+    response::{
+        sse::{Event, KeepAlive, Sse},
+        IntoResponse, Response,
+    },
+    Json,
+};
+use harness_protocol::methods::RpcRequest;
+use serde_json::json;
+use std::sync::Arc;
+
+pub(crate) async fn health_check(State(state): State<Arc<AppState>>) -> Json<serde_json::Value> {
+    let count = state.core.tasks.count();
+    Json(json!({"status": "ok", "tasks": count}))
+}
+
+/// GET /projects — per-project queue stats alongside the global queue summary.
+pub(crate) async fn project_queue_stats(
+    State(state): State<Arc<AppState>>,
+) -> Json<serde_json::Value> {
+    let tq = &state.concurrency.task_queue;
+    let projects: serde_json::Map<String, serde_json::Value> = tq
+        .all_project_stats()
+        .into_iter()
+        .map(|(id, s)| {
+            (
+                id,
+                json!({
+                    "running": s.running,
+                    "queued": s.queued,
+                    "limit": s.limit,
+                }),
+            )
+        })
+        .collect();
+    Json(json!({
+        "global": {
+            "running": tq.running_count(),
+            "queued": tq.queued_count(),
+            "limit": tq.global_limit(),
+        },
+        "projects": projects,
+    }))
+}
+
+pub(crate) async fn handle_rpc(
+    State(state): State<Arc<AppState>>,
+    Json(req): Json<RpcRequest>,
+) -> Response {
+    match crate::router::handle_request(&state, req).await {
+        Some(resp) => (StatusCode::OK, Json(resp)).into_response(),
+        None => StatusCode::NO_CONTENT.into_response(),
+    }
+}
+
+pub(crate) async fn github_webhook(
+    State(state): State<Arc<AppState>>,
+    headers: HeaderMap,
+    body: Bytes,
+) -> (StatusCode, Json<serde_json::Value>) {
+    let secret = match state
+        .core
+        .server
+        .config
+        .server
+        .github_webhook_secret
+        .as_deref()
+    {
+        Some("") => {
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "invalid server.github_webhook_secret configuration"})),
+            )
+        }
+        Some(secret) => secret,
+        None => {
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "missing server.github_webhook_secret configuration"})),
+            )
+        }
+    };
+    let signature = match headers
+        .get("x-hub-signature-256")
+        .and_then(|value| value.to_str().ok())
+    {
+        Some(signature) => signature,
+        None => {
+            return (
+                StatusCode::UNAUTHORIZED,
+                Json(json!({"error": "missing header x-hub-signature-256"})),
+            )
+        }
+    };
+    if !crate::webhook::verify_github_signature(secret, signature, body.as_ref()) {
+        return (
+            StatusCode::UNAUTHORIZED,
+            Json(json!({"error": "invalid webhook signature"})),
+        );
+    }
+
+    let event = match headers
+        .get("x-github-event")
+        .and_then(|value| value.to_str().ok())
+    {
+        Some(event) => event,
+        None => {
+            return (
+                StatusCode::BAD_REQUEST,
+                Json(json!({"error": "missing header x-github-event"})),
+            )
+        }
+    };
+    if !crate::webhook::is_valid_github_event_name(event) {
+        return (
+            StatusCode::BAD_REQUEST,
+            Json(json!({"error": "invalid header x-github-event"})),
+        );
+    }
+
+    let (request, reason) =
+        match crate::webhook::parse_github_webhook_task_request(event, body.as_ref()) {
+            Ok(parsed) => parsed,
+            Err(error) => return (StatusCode::BAD_REQUEST, Json(json!({ "error": error }))),
+        };
+
+    let Some(mut req) = request else {
+        return (
+            StatusCode::OK,
+            Json(json!({
+                "status": "ignored",
+                "reason": reason,
+            })),
+        );
+    };
+
+    if req.project.is_none() {
+        req.project = Some(state.core.project_root.clone());
+    }
+
+    match super::task_routes::enqueue_task(&state, req).await {
+        Ok(task_id) => (
+            StatusCode::ACCEPTED,
+            Json(json!({
+                "status": "accepted",
+                "reason": reason,
+                "task_id": task_id.0,
+            })),
+        ),
+        Err(crate::services::execution::EnqueueTaskError::BadRequest(error)) => {
+            (StatusCode::BAD_REQUEST, Json(json!({ "error": error })))
+        }
+        Err(crate::services::execution::EnqueueTaskError::Internal(error)) => (
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json(json!({ "error": error })),
+        ),
+    }
+}
+
+pub(crate) async fn list_tasks(State(state): State<Arc<AppState>>) -> Response {
+    match state.core.tasks.list_all_summaries_with_terminal().await {
+        Ok(summaries) => Json(summaries).into_response(),
+        Err(e) => {
+            tracing::error!("list_tasks: database error: {e}");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response()
+        }
+    }
+}
+
+pub(crate) async fn get_task(
+    State(state): State<Arc<AppState>>,
+    Path(id): Path<String>,
+) -> Response {
+    match state
+        .core
+        .tasks
+        .get_with_db_fallback(&harness_core::types::TaskId(id))
+        .await
+    {
+        Ok(Some(task)) => Json(task).into_response(),
+        Ok(None) => (
+            StatusCode::NOT_FOUND,
+            Json(json!({"error": "task not found"})),
+        )
+            .into_response(),
+        Err(e) => {
+            tracing::error!("get_task: database error: {e}");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response()
+        }
+    }
+}
+
+/// GET /tasks/{id}/artifacts — all persisted artifacts for a task.
+pub(crate) async fn get_task_artifacts(
+    State(state): State<Arc<AppState>>,
+    Path(id): Path<String>,
+) -> Response {
+    let task_id = harness_core::types::TaskId(id);
+    match state.core.tasks.get_with_db_fallback(&task_id).await {
+        Ok(None) => {
+            return (
+                StatusCode::NOT_FOUND,
+                Json(json!({"error": "task not found"})),
+            )
+                .into_response();
+        }
+        Err(e) => {
+            tracing::error!("get_task_artifacts: database error: {e}");
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response();
+        }
+        Ok(Some(_)) => {}
+    }
+    match state.core.tasks.list_artifacts(&task_id).await {
+        Ok(artifacts) => Json(artifacts).into_response(),
+        Err(e) => {
+            tracing::error!("get_task_artifacts: list artifacts error: {e}");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response()
+        }
+    }
+}
+
+/// GET /tasks/{id}/stream — real-time SSE stream of agent execution events.
+///
+/// Subscribes to the task's broadcast channel and forwards each [`StreamItem`]
+/// as a JSON-encoded SSE data event. The stream ends when the task completes
+/// (channel closed). If the receiver lags, a synthetic "lag" event is emitted
+/// noting how many events were dropped; streaming then continues.
+pub(crate) async fn stream_task_sse(
+    State(state): State<Arc<AppState>>,
+    Path(id): Path<String>,
+) -> Response {
+    let task_id = harness_core::types::TaskId(id);
+
+    let rx = match state.core.tasks.subscribe_task_stream(&task_id) {
+        Some(rx) => rx,
+        None => {
+            match state.core.tasks.get_with_db_fallback(&task_id).await {
+                Ok(None) => {
+                    return (
+                        StatusCode::NOT_FOUND,
+                        Json(json!({"error": "task not found"})),
+                    )
+                        .into_response();
+                }
+                Err(e) => {
+                    tracing::error!("stream_task_sse: database error: {e}");
+                    return (
+                        StatusCode::INTERNAL_SERVER_ERROR,
+                        Json(json!({"error": "internal server error"})),
+                    )
+                        .into_response();
+                }
+                Ok(Some(_)) => {}
+            }
+            // Task exists but stream already closed (task completed before client connected).
+            let stream = futures::stream::empty::<Result<Event, std::convert::Infallible>>();
+            return Sse::new(stream).into_response();
+        }
+    };
+
+    let stream = futures::stream::unfold(rx, |mut rx| async move {
+        match rx.recv().await {
+            Ok(item) => {
+                let data = match serde_json::to_string(&item) {
+                    Ok(s) => s,
+                    Err(e) => {
+                        tracing::warn!("sse: failed to serialize event: {e}");
+                        String::new()
+                    }
+                };
+                Some((
+                    Ok::<Event, std::convert::Infallible>(Event::default().data(data)),
+                    rx,
+                ))
+            }
+            Err(tokio::sync::broadcast::error::RecvError::Closed) => None,
+            Err(tokio::sync::broadcast::error::RecvError::Lagged(n)) => {
+                let event = Event::default()
+                    .event("lag")
+                    .data(format!("dropped {n} events due to slow consumer"));
+                Some((Ok(event), rx))
+            }
+        }
+    });
+
+    // Send a heartbeat comment every 30 s so reverse proxies (nginx default
+    // 60 s idle timeout) don't drop the connection while the agent is silent.
+    Sse::new(stream)
+        .keep_alive(
+            KeepAlive::new()
+                .interval(std::time::Duration::from_secs(30))
+                .text("heartbeat"),
+        )
+        .into_response()
+}
