+++ crates/harness-server/src/http/helpers.rs
@@ -1,0 +1,96 @@
+use std::sync::Arc;
+
+/// Post a comment to a GitHub PR via the Issues API.
+pub(crate) async fn post_review_bot_comment(
+    owner: &str,
+    repo: &str,
+    pr_number: u64,
+    body: &str,
+    github_token: &str,
+) -> anyhow::Result<()> {
+    let url = format!("https://api.github.com/repos/{owner}/{repo}/issues/{pr_number}/comments");
+    let client = reqwest::Client::new();
+    let resp = client
+        .post(&url)
+        .header("Authorization", format!("Bearer {github_token}"))
+        .header("Accept", "application/vnd.github+json")
+        .header("X-GitHub-Api-Version", "2022-11-28")
+        .header("User-Agent", "harness-bot")
+        .json(&serde_json::json!({ "body": body }))
+        .send()
+        .await?;
+    if !resp.status().is_success() {
+        let status = resp.status();
+        let text = resp.text().await.unwrap_or_default();
+        anyhow::bail!("GitHub API returned {status}: {text}");
+    }
+    Ok(())
+}
+
+/// Resolve the reviewer agent for independent agent review.
+///
+/// 1. If `config.reviewer_agent` is set and differs from implementor, use it.
+/// 2. Otherwise, auto-select the first registered agent that isn't the implementor.
+/// 3. If none found, return None (agent review will be skipped).
+pub(crate) fn resolve_reviewer(
+    registry: &harness_agents::registry::AgentRegistry,
+    config: &harness_core::config::agents::AgentReviewConfig,
+    implementor_name: &str,
+) -> (
+    Option<Arc<dyn harness_core::agent::CodeAgent>>,
+    harness_core::config::agents::AgentReviewConfig,
+) {
+    if !config.enabled {
+        return (None, config.clone());
+    }
+
+    // Explicit reviewer
+    if !config.reviewer_agent.is_empty() {
+        if config.reviewer_agent == implementor_name {
+            tracing::warn!(
+                "agents.review.reviewer_agent == implementor '{}', skipping agent review",
+                implementor_name
+            );
+            return (None, config.clone());
+        }
+        if let Some(agent) = registry.get(&config.reviewer_agent) {
+            return (Some(agent), config.clone());
+        }
+        tracing::warn!(
+            "agents.review.reviewer_agent '{}' not registered, skipping agent review",
+            config.reviewer_agent
+        );
+        return (None, config.clone());
+    }
+
+    // Auto-select: first agent != implementor
+    for name in registry.list() {
+        if name != implementor_name {
+            if let Some(agent) = registry.get(name) {
+                return (Some(agent), config.clone());
+            }
+        }
+    }
+
+    (None, config.clone())
+}
+
+/// Extract the PR number from a GitHub PR URL.
+///
+/// Handles:
+/// - `.../pull/42`
+/// - `.../pull/42/files`
+/// - `.../pull/42#discussion_r...`
+pub(crate) fn parse_pr_num_from_url(url: &str) -> Option<u64> {
+    // Strip fragment first, then query string
+    let url = url.split('#').next().unwrap_or(url);
+    let url = url.split('?').next().unwrap_or(url);
+    // Walk path segments looking for "pull", then parse the segment that follows
+    let mut parts = url.split('/');
+    while let Some(seg) = parts.next() {
+        if seg == "pull" {
+            return parts.next()?.parse::<u64>().ok();
+        }
+    }
+    None
+}
