+++ crates/harness-server/src/http/intake_auth.rs
@@ -1,0 +1,290 @@
+use super::types::AppState;
+use crate::task_runner;
+use axum::{
+    body::Bytes,
+    extract::State,
+    http::{HeaderMap, StatusCode},
+    Json,
+};
+use serde_json::json;
+use std::sync::Arc;
+
+/// GET /api/intake — current status of all intake channels and recent dispatches.
+pub(crate) async fn intake_status(State(state): State<Arc<AppState>>) -> Json<serde_json::Value> {
+    let intake_config = &state.core.server.config.intake;
+    let all_tasks = state.core.tasks.list_all();
+
+    let github_active: u64 = all_tasks
+        .iter()
+        .filter(|t| {
+            t.source.as_deref() == Some("github")
+                && !matches!(
+                    t.status,
+                    task_runner::TaskStatus::Done | task_runner::TaskStatus::Failed
+                )
+        })
+        .count() as u64;
+
+    let feishu_active: u64 = all_tasks
+        .iter()
+        .filter(|t| {
+            t.source.as_deref() == Some("feishu")
+                && !matches!(
+                    t.status,
+                    task_runner::TaskStatus::Done | task_runner::TaskStatus::Failed
+                )
+        })
+        .count() as u64;
+
+    let dashboard_active: u64 = all_tasks
+        .iter()
+        .filter(|t| {
+            (t.source.as_deref() == Some("dashboard") || t.source.is_none())
+                && !matches!(
+                    t.status,
+                    task_runner::TaskStatus::Done | task_runner::TaskStatus::Failed
+                )
+        })
+        .count() as u64;
+
+    let github_channel = json!({
+        "name": "github",
+        "enabled": intake_config.github.as_ref().map(|c| c.enabled).unwrap_or(false),
+        "repo": intake_config.github.as_ref().map(|c| c.repo.as_str()).unwrap_or(""),
+        "active": github_active,
+    });
+
+    let feishu_channel = json!({
+        "name": "feishu",
+        "enabled": state.intake.feishu_intake.is_some(),
+        "keyword": intake_config.feishu.as_ref().map(|c| c.trigger_keyword.as_str()).unwrap_or(""),
+        "active": feishu_active,
+    });
+
+    let dashboard_channel = json!({
+        "name": "dashboard",
+        "enabled": true,
+        "active": dashboard_active,
+    });
+
+    let mut recent_dispatches: Vec<serde_json::Value> = all_tasks
+        .iter()
+        .filter(|t| t.source.is_some())
+        .map(|t| {
+            json!({
+                "source": t.source,
+                "external_id": t.external_id,
+                "task_id": t.id.0,
+                "status": serde_json::to_value(&t.status).unwrap_or(json!("unknown")),
+                "pr_url": t.pr_url,
+            })
+        })
+        .collect();
+    recent_dispatches.truncate(10);
+
+    Json(json!({
+        "channels": [github_channel, feishu_channel, dashboard_channel],
+        "recent_dispatches": recent_dispatches,
+    }))
+}
+
+/// Request body for `POST /signals`.
+#[derive(serde::Deserialize)]
+pub(crate) struct IngestSignalRequest {
+    source: String,
+    #[serde(default)]
+    severity: Option<harness_core::types::Severity>,
+    payload: serde_json::Value,
+}
+
+/// Infer severity from a GitHub webhook payload: CI failure → High, changes_requested → Medium.
+fn infer_github_severity(payload: &serde_json::Value) -> Option<harness_core::types::Severity> {
+    if let Some(obj) = payload.as_object() {
+        // check_run completed with failure
+        if let (Some(action), Some(check_run)) = (
+            obj.get("action").and_then(|v| v.as_str()),
+            obj.get("check_run"),
+        ) {
+            if action == "completed" {
+                let conclusion = check_run
+                    .get("conclusion")
+                    .and_then(|v| v.as_str())
+                    .unwrap_or("");
+                if conclusion == "failure" {
+                    return Some(harness_core::types::Severity::High);
+                }
+            }
+        }
+        // pull_request_review with changes_requested
+        if let Some(review) = obj.get("review") {
+            let state = review.get("state").and_then(|v| v.as_str()).unwrap_or("");
+            if state.eq_ignore_ascii_case("changes_requested") {
+                return Some(harness_core::types::Severity::Medium);
+            }
+        }
+    }
+    None
+}
+
+/// POST /signals — ingest an external signal (CI failure, review feedback, etc.).
+///
+/// Validates the `x-hub-signature-256` HMAC-SHA256 header using the configured
+/// `server.github_webhook_secret`. Rate-limited to 100 requests per source per minute.
+pub(crate) async fn ingest_signal(
+    State(state): State<Arc<AppState>>,
+    headers: HeaderMap,
+    body: Bytes,
+) -> (StatusCode, Json<serde_json::Value>) {
+    let secret = match state
+        .core
+        .server
+        .config
+        .server
+        .github_webhook_secret
+        .as_deref()
+    {
+        Some("") | None => {
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "server.github_webhook_secret not configured"})),
+            )
+        }
+        Some(s) => s,
+    };
+
+    let signature = match headers
+        .get("x-hub-signature-256")
+        .and_then(|v| v.to_str().ok())
+    {
+        Some(sig) => sig,
+        None => {
+            return (
+                StatusCode::UNAUTHORIZED,
+                Json(json!({"error": "missing header x-hub-signature-256"})),
+            )
+        }
+    };
+
+    if !crate::webhook::verify_github_signature(secret, signature, body.as_ref()) {
+        return (
+            StatusCode::UNAUTHORIZED,
+            Json(json!({"error": "invalid webhook signature"})),
+        );
+    }
+
+    let req: IngestSignalRequest = match serde_json::from_slice(&body) {
+        Ok(r) => r,
+        Err(e) => {
+            return (
+                StatusCode::BAD_REQUEST,
+                Json(json!({"error": format!("invalid payload: {e}")})),
+            )
+        }
+    };
+
+    if !state
+        .observability
+        .signal_rate_limiter
+        .check_and_increment(&req.source)
+    {
+        return (
+            StatusCode::TOO_MANY_REQUESTS,
+            Json(json!({"error": "rate limit exceeded: max 100 signals per minute per source"})),
+        );
+    }
+
+    let severity = req.severity.unwrap_or_else(|| {
+        if req.source == "github" {
+            infer_github_severity(&req.payload).unwrap_or(harness_core::types::Severity::Low)
+        } else {
+            harness_core::types::Severity::Low
+        }
+    });
+
+    let signal =
+        harness_core::types::ExternalSignal::new(req.source.clone(), severity, req.payload.clone());
+    let signal_id = signal.id.clone();
+
+    if let Err(e) = state.observability.events.log_external_signal(&signal) {
+        tracing::error!(source = %req.source, "failed to store external signal: {e}");
+        return (
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json(json!({"error": "failed to store signal"})),
+        );
+    }
+
+    tracing::info!(
+        source = %req.source,
+        severity = ?severity,
+        signal_id = %signal_id,
+        "external signal ingested"
+    );
+
+    (
+        StatusCode::OK,
+        Json(json!({"status": "accepted", "id": signal_id.as_str()})),
+    )
+}
+
+#[derive(serde::Deserialize)]
+pub(crate) struct PasswordResetRequest {
+    email: String,
+}
+
+/// POST /auth/reset-password — initiate a password reset.
+///
+/// Rate-limited per email address to prevent enumeration and brute-force.
+/// Always returns a generic success response regardless of whether the email
+/// exists, to avoid leaking account information.
+pub(crate) async fn password_reset(
+    State(state): State<Arc<AppState>>,
+    Json(req): Json<PasswordResetRequest>,
+) -> (StatusCode, Json<serde_json::Value>) {
+    let email = req.email.trim().to_lowercase();
+    if email.is_empty() {
+        return (
+            StatusCode::BAD_REQUEST,
+            Json(json!({"error": "email is required"})),
+        );
+    }
+
+    let limit = state
+        .core
+        .server
+        .config
+        .server
+        .password_reset_rate_limit_per_hour;
+
+    if !state
+        .observability
+        .password_reset_rate_limiter
+        .check_and_increment(&email)
+    {
+        return (
+            StatusCode::TOO_MANY_REQUESTS,
+            Json(json!({
+                "error": format!(
+                    "rate limit exceeded: max {} password reset requests per hour",
+                    limit
+                )
+            })),
+        );
+    }
+
+    tracing::info!(
+        email_hash = %format!("{:x}", {
+            use std::hash::{Hash, Hasher};
+            let mut h = std::collections::hash_map::DefaultHasher::new();
+            email.hash(&mut h);
+            h.finish()
+        }),
+        "password reset requested"
+    );
+
+    (
+        StatusCode::OK,
+        Json(
+            json!({"status": "ok", "message": "If that email is registered, a reset link has been sent."}),
+        ),
+    )
+}
