+++ crates/harness-server/src/http/tests.rs
@@ -1,17 +1,26 @@
 use super::*;
 use async_trait::async_trait;
 use axum::body::Body;
-use axum::http::Request;
+use axum::extract::DefaultBodyLimit;
+use axum::http::{Request, StatusCode};
+use axum::routing::{get, post};
+use axum::Router;
 use harness_core::{
     agent::AgentRequest, agent::AgentResponse, agent::CodeAgent, agent::StreamItem,
     types::Capability, types::TokenUsage,
 };
 use hmac::{Hmac, Mac};
 use sha2::Sha256;
+use std::sync::atomic::{AtomicBool, AtomicU64};
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::Mutex;
 use tower::ServiceExt;
+
+use super::handlers::github_webhook;
+use super::handlers::{get_task, health_check, list_tasks};
+use super::intake_auth::intake_status;
+use crate::task_runner;
 
 struct CapturingAgent {
     prompts: Mutex<Vec<String>>,
